Trace volcano plot for computed CIS data.

Traces a volcano plot for IS frequency and CIS results.

CIS_volcano_plot(
  x,
  onco_db_file = system.file("extdata", "201806_uniprot-Proto-oncogene.tsv.xz", package
    = "ISAnalytics"),
  tumor_suppressors_db_file = system.file("extdata",
    "201806_uniprot-Tumor-suppressor.tsv.xz", package = "ISAnalytics"),
  species = "human",
  known_onco = known_clinical_oncogenes(),
  suspicious_genes = clinical_relevant_suspicious_genes(),
  significance_threshold = 0.05,
  annotation_threshold_ontots = 0.1,
  highlight_genes = NULL,
  title_prefix = NULL,
  return_df = FALSE
)

Arguments

x
Either a simple integration matrix or a data frame resulting from the call to CIS_grubbs with add_standard_padjust = TRUE

onco_db_file
Uniprot file for proto-oncogenes (see details)

tumor_suppressors_db_file
Uniprot file for tumor-suppressor genes

species
One between "human", "mouse" and "all"

known_onco
Data frame with known oncogenes. See details.

suspicious_genes
Data frame with clinical relevant suspicious genes. See details.

significance_threshold
The significance threshold

annotation_threshold_ontots
Value above which genes are annotated

highlight_genes
Either NULL or a character vector of genes to be highlighted in the plot even if they're not above the threshold

title_prefix
A string to be displayed in the title - usually the project name and other characterizing info

return_df
Return the data frame used to generate the plot? This can be useful if the user wants to manually modify the plot with ggplot2. If TRUE the function returns a list containing both the plot and the data frame.

Value

A plot or a list containing a plot and a data frame

Details

Input data frame

Users can supply as x either a simple integration matrix or a data frame resulting from the call to CIS_grubbs with add_standard_padjust = TRUE. In the first case an internal call to the function CIS_grubbs is performed.

Oncogene and tumor suppressor genes files

These files are included in the package for user convenience and are simply UniProt files with gene annotations for human and mouse. For more details on how this files were generated use the help ?filename function.

Known oncogenes

The default values are contained in a data frame exported by this package, it can be accessed by doing:

head(known_clinical_oncogenes())

## # A tibble: 5 x 2
##   GeneName KnownClonalExpansion
##   <chr>    <lgl>               
## 1 MECOM    TRUE                
## 2 CCND2    TRUE                
## 3 TAL1     TRUE                
## 4 LMO2     TRUE                
## 5 HMGA2    TRUE

If the user wants to change this parameter the input data frame must preserve the column structure. The same goes for the suspicious_genes parameter (DOIReference column is optional):

head(clinical_relevant_suspicious_genes())

## # A tibble: 6 x 3
##   GeneName ClinicalRelevance DOIReference                                
##   <chr>    <lgl>             <chr>                                       
## 1 DNMT3A   TRUE              https://doi.org/10.1182/blood-2018-01-829937
## 2 TET2     TRUE              https://doi.org/10.1182/blood-2018-01-829937
## 3 ASXL1    TRUE              https://doi.org/10.1182/blood-2018-01-829937
## 4 JAK2     TRUE              https://doi.org/10.1182/blood-2018-01-829937
## 5 CBL      TRUE              https://doi.org/10.1182/blood-2018-01-829937
## 6 TP53     TRUE              https://doi.org/10.1182/blood-2018-01-829937

Examples

op <- options(ISAnalytics.widgets = FALSE)

path_AF <- system.file("extdata", "ex_association_file.tsv",
    package = "ISAnalytics"
)
root_correct <- system.file("extdata", "fs.zip",
    package = "ISAnalytics"
)
root_correct <- unzip_file_system(root_correct, "fs")

matrices <- import_parallel_Vispa2Matrices_auto(
    association_file = path_AF, root = root_correct,
    quantification_type = c("seqCount", "fragmentEstimate"),
    matrix_type = "annotated", workers = 2, patterns = NULL,
    matching_opt = "ANY",
    dates_format = "dmy"
)

#> *** Association file import summary ***
#> ℹ For detailed report please set option 'ISAnalytics.widgets' to TRUE
#> * Parsing problems detected: TRUE
#> * Date parsing problems: TRUE
#> * Column problems detected: TRUE
#> * NAs found in important columns: TRUE
#> * File system alignment: no problems detected

#> [1] "--- REPORT: FILES IMPORTED ---"
#> # A tibble: 8 x 5
#>   ProjectID   concatenatePoolIDSeqRun Quantification_type
#>   <chr>       <chr>                   <chr>              
#> 1 CLOEXP      POOL6-1                 seqCount           
#> 2 PROJECT1100 ABX-LR-PL5-POOL14-1     seqCount           
#> 3 PROJECT1100 ABX-LR-PL6-POOL15-1     seqCount           
#> 4 PROJECT1101 ABY-LR-PL4-POOL54-2     seqCount           
#> 5 CLOEXP      POOL6-1                 fragmentEstimate   
#> 6 PROJECT1100 ABX-LR-PL5-POOL14-1     fragmentEstimate   
#> 7 PROJECT1100 ABX-LR-PL6-POOL15-1     fragmentEstimate   
#> 8 PROJECT1101 ABY-LR-PL4-POOL54-2     fragmentEstimate   
#>   Files_chosen                                                                  
#>   <fs::path>                                                                    
#> 1 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpWPuXBB/fs/VA/CLOEXP/quan…
#> 2 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpWPuXBB/fs/PROJECT1100/qu…
#> 3 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpWPuXBB/fs/PROJECT1100/qu…
#> 4 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpWPuXBB/fs/PROJECT1101/qu…
#> 5 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpWPuXBB/fs/VA/CLOEXP/quan…
#> 6 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpWPuXBB/fs/PROJECT1100/qu…
#> 7 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpWPuXBB/fs/PROJECT1100/qu…
#> 8 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpWPuXBB/fs/PROJECT1101/qu…
#>   Imported
#>   <lgl>   
#> 1 TRUE    
#> 2 TRUE    
#> 3 TRUE    
#> 4 TRUE    
#> 5 TRUE    
#> 6 TRUE    
#> 7 TRUE    
#> 8 TRUE    
#> [1] "--- SUMMARY OF FILES CHOSEN FOR IMPORT ---"
#> # A tibble: 8 x 4
#>   ProjectID   concatenatePoolIDSeqRun Quantification_type
#>   <chr>       <chr>                   <chr>              
#> 1 CLOEXP      POOL6-1                 seqCount           
#> 2 CLOEXP      POOL6-1                 fragmentEstimate   
#> 3 PROJECT1100 ABX-LR-PL5-POOL14-1     seqCount           
#> 4 PROJECT1100 ABX-LR-PL5-POOL14-1     fragmentEstimate   
#> 5 PROJECT1100 ABX-LR-PL6-POOL15-1     seqCount           
#> 6 PROJECT1100 ABX-LR-PL6-POOL15-1     fragmentEstimate   
#> 7 PROJECT1101 ABY-LR-PL4-POOL54-2     seqCount           
#> 8 PROJECT1101 ABY-LR-PL4-POOL54-2     fragmentEstimate   
#>   Files_chosen                                                                  
#>   <fs::path>                                                                    
#> 1 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpWPuXBB/fs/VA/CLOEXP/quan…
#> 2 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpWPuXBB/fs/VA/CLOEXP/quan…
#> 3 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpWPuXBB/fs/PROJECT1100/qu…
#> 4 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpWPuXBB/fs/PROJECT1100/qu…
#> 5 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpWPuXBB/fs/PROJECT1100/qu…
#> 6 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpWPuXBB/fs/PROJECT1100/qu…
#> 7 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpWPuXBB/fs/PROJECT1101/qu…
#> 8 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpWPuXBB/fs/PROJECT1101/qu…
#> [1] "--- INTEGRATION MATRICES FOUND REPORT ---"
#> # A tibble: 8 x 5
#>   ProjectID   concatenatePoolIDSeqRun Anomalies Quantification_type
#>   <chr>       <chr>                   <lgl>     <chr>              
#> 1 CLOEXP      POOL6-1                 FALSE     seqCount           
#> 2 CLOEXP      POOL6-1                 FALSE     fragmentEstimate   
#> 3 PROJECT1100 ABX-LR-PL5-POOL14-1     FALSE     seqCount           
#> 4 PROJECT1100 ABX-LR-PL5-POOL14-1     FALSE     fragmentEstimate   
#> 5 PROJECT1100 ABX-LR-PL6-POOL15-1     FALSE     seqCount           
#> 6 PROJECT1100 ABX-LR-PL6-POOL15-1     FALSE     fragmentEstimate   
#> 7 PROJECT1101 ABY-LR-PL4-POOL54-2     FALSE     seqCount           
#> 8 PROJECT1101 ABY-LR-PL4-POOL54-2     FALSE     fragmentEstimate   
#>   Files_found                                                                   
#>   <fs::path>                                                                    
#> 1 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpWPuXBB/fs/VA/CLOEXP/quan…
#> 2 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpWPuXBB/fs/VA/CLOEXP/quan…
#> 3 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpWPuXBB/fs/PROJECT1100/qu…
#> 4 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpWPuXBB/fs/PROJECT1100/qu…
#> 5 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpWPuXBB/fs/PROJECT1100/qu…
#> 6 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpWPuXBB/fs/PROJECT1100/qu…
#> 7 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpWPuXBB/fs/PROJECT1101/qu…
#> 8 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpWPuXBB/fs/PROJECT1101/qu…

cis <- CIS_grubbs(matrices)
plot <- CIS_volcano_plot(cis)

#> Loading annotated genes -  species selected: Homo sapiens (Human)

#> Loading annotated genes -  done

options(op)