ISAnalytics: Analyze gene therapy vector insertion sites data identified from genomics next generation sequencing reads for clonal tracking studies

In gene therapy, stem cells are modified using viral vectors to deliver the therapeutic transgene and replace functional properties since the genetic modification is stable and inherited in all cell progeny. The retrieval and mapping of the sequences flanking the virus-host DNA junctions allows the identification of insertion sites (IS), essential for monitoring the evolution of genetically modified cells in vivo. A comprehensive toolkit for the analysis of IS is required to foster clonal trackign studies and supporting the assessment of safety and long term efficacy in vivo. This package is aimed at (1) supporting automation of IS workflow, (2) performing base and advance analysis for IS tracking (clonal abundance, clonal expansions and statistics for insertional mutagenesis, etc.), (3) providing basic biology insights of transduced stem cells in vivo.

Useful resources

VISPA2: A Scalable Pipeline for High-Throughput Identification and Annotation of Vector Integration Sites

ISAnalytics function families

Import functions:

import_single_Vispa2Matrix

import_association_file

import_parallel_Vispa2Matrices_interactive

import_parallel_Vispa2Matrices_auto

Aggregation functions:

aggregate_metadata

aggregate_values_by_key

Collision removal functions:

remove_collisions

realign_after_collisions

Recalibration functions:

compute_near_integrations

Analysis functions:

compute_abundance

Utility functions:

generate_blank_association_file

generate_Vispa2_launch_AF

unzip_file_system

Vignettes

vignette("How to use import functions", package = "ISAnalytics")

vignette("Collision removal functionality", package = "ISAnalytics")

vignette("Working with aggregate functions", package = "ISAnalytics")