obtain a single integration matrix from individual quantification matrices.

Takes a list of integration matrices referring to different quantification types and merges them in a single data frame that has multiple value columns, each renamed according to their quantification type of reference.

Usage

comparison_matrix(
  x,
  fragmentEstimate = "fragmentEstimate",
  seqCount = "seqCount",
  barcodeCount = "barcodeCount",
  cellCount = "cellCount",
  ShsCount = "ShsCount"
)

Arguments

x

A named list of integration matrices, ideally obtained via import_parallel_Vispa2Matrices_interactive or import_parallel_Vispa2Matrices_auto. Names must be quantification types.

fragmentEstimate

The name of the output column for fragment estimate values

seqCount

The name of the output column for sequence count values

barcodeCount

The name of the output column for barcode count values

cellCount

The name of the output column for cell count values

ShsCount

The name of the output column for Shs count values

Value

A tibble

See also

Examples

fs_path <- system.file("extdata", "fs.zip", package = "ISAnalytics")
fs <- unzip_file_system(fs_path, "fs")
af_path <- system.file("extdata", "asso.file.tsv.gz",
    package = "ISAnalytics"
)
af <- import_association_file(af_path,
    root = fs,
    import_iss = FALSE,
    report_path = NULL
)
matrices <- import_parallel_Vispa2Matrices(af,
    c("seqCount", "fragmentEstimate"),
    mode = "AUTO", report_path = NULL, multi_quant_matrix = FALSE
)
multi_quant <- comparison_matrix(matrices)
head(multi_quant)
#> # A tibble: 6 × 8
#>   chr   integration_locus strand GeneName     GeneStrand CompleteAmplificationID
#>   <chr>             <int> <chr>  <chr>        <chr>      <fct>                  
#> 1 16             68164148 +      NFATC3       +          PJ01_POOL01_LTR75LC38_…
#> 2 4             129390130 +      LOC100507487 +          PJ01_POOL01_LTR75LC38_…
#> 3 5              84009671 -      EDIL3        -          PJ01_POOL01_LTR75LC38_…
#> 4 12             54635693 -      CBX5         -          PJ01_POOL01_LTR75LC38_…
#> 5 2             181930711 +      UBE2E3       +          PJ01_POOL01_LTR75LC38_…
#> 6 20             35920986 +      MANBAL       +          PJ01_POOL01_LTR75LC38_…
#> # … with 2 more variables: fragmentEstimate <dbl>, seqCount <int>