Computes the abundance for every integration event in the input data frame.

Abundance is obtained for every integration event by calculating the ratio between the single value and the total value for the given group.

compute_abundance(
  x,
  columns = "Value",
  percentage = TRUE,
  key = "CompleteAmplificationID",
  keep_totals = FALSE
)

Arguments

x
An integration matrix - aka a data frame that includes the mandatory_IS_vars() as columns. The matrix can either be aggregated (via aggregate_values_by_key()) or not.

columns
A character vector of column names to process, must be numeric or integer columns

percentage
Add abundance as percentage?

key
The key to group by when calculating totals

keep_totals
A value between TRUE, FALSE or df. If TRUE, the intermediate totals for each group will be kept in the output data frame as a dedicated column with a trailing "_tot". If FALSE, totals won't be included in the output data frame. If df, the totals are returned to the user as a separate data frame, together with the abundance data frame.

Value

Either a single data frame with computed abundance values or a list of 2 data frames (abundance_df, quant_totals)

Details

Abundance will be computed upon the user selected columns in the columns parameter. For each column a corresponding relative abundance column (and optionally a percentage abundance column) will be produced.

See also

Examples

path <- system.file("extdata", "ex_annotated_ISMatrix.tsv.xz",
    package = "ISAnalytics"
)
matrix <- import_single_Vispa2Matrix(path)

#> Warning: compression format not supported by fread
#> ℹ File will be read using readr

#> Reading file...
#> ℹ Mode: classic

#> *** File info *** 
#> * --- Matrix type: NEW
#> * --- Annotated: TRUE
#> * --- Dimensions: 50 x 10
#> * --- Read mode:  classic

#> Reshaping...

#>   |                                                                              |                                                                      |   0%  |                                                                              |===                                                                   |   5%  |                                                                              |======                                                                |   9%  |                                                                              |==========                                                            |  14%  |                                                                              |=============                                                         |  18%  |                                                                              |================                                                      |  23%  |                                                                              |===================                                                   |  27%  |                                                                              |======================                                                |  32%  |                                                                              |=========================                                             |  36%  |                                                                              |=============================                                         |  41%  |                                                                              |================================                                      |  45%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================                                |  55%  |                                                                              |=========================================                             |  59%  |                                                                              |=============================================                         |  64%  |                                                                              |================================================                      |  68%  |                                                                              |===================================================                   |  73%  |                                                                              |======================================================                |  77%  |                                                                              |=========================================================             |  82%  |                                                                              |============================================================          |  86%  |                                                                              |================================================================      |  91%  |                                                                              |===================================================================   |  95%  |                                                                              |======================================================================| 100%
#> 

#> Done!

# Simple integration matrix - grouping by CompleteAmplificationID
abundance1 <- compute_abundance(matrix)
abundance1

#>     chr integration_locus strand  GeneName GeneStrand CompleteAmplificationID
#>  1:  20          39085641      +      MAFB          -                   exp_2
#>  2:  20          39085641      +      MAFB          -                   exp_5
#>  3:   7         155446768      +     RBM33          +                   exp_3
#>  4:  19           4049601      +    ZBTB7A          -                   exp_1
#>  5:  19           4049601      +    ZBTB7A          -                   exp_2
#>  6:  19          12278742      -    ZNF136          +                   exp_3
#>  7:  19           4049601      +    ZBTB7A          -                   exp_3
#>  8:  19          12278742      -    ZNF136          +                   exp_5
#>  9:   6           1865825      -      GMDS          -                   exp_4
#> 10:   6          70219927      -    ADGRB3          +                   exp_5
#> 11:  17          79831324      +   ARHGDIA          -                   exp_1
#> 12:  17           7113678      +      DLG4          -                   exp_3
#> 13:  17          56480574      +     RNF43          -                   exp_4
#> 14:  13         115089373      +    CHAMP1          +                   exp_4
#> 15:   4         127380734      -   MIR2054          +                   exp_3
#> 16:   4          46971398      -    GABRA4          -                   exp_4
#> 17:  11          57153892      +      PRG2          -                   exp_1
#> 18:  11          57153892      +      PRG2          -                   exp_2
#> 19:  11          57153892      +      PRG2          -                   exp_4
#> 20:  11          57153892      +      PRG2          -                   exp_5
#> 21:  15          56979473      -   ZNF280D          -                   exp_3
#> 22:  15          76209162      -    FBXO22          +                   exp_4
#> 23:  15          56979473      -   ZNF280D          -                   exp_5
#> 24:  10          99253610      +     MMS19          -                   exp_2
#> 25:  10          38314134      -    ZNF33A          +                   exp_4
#> 26:  10          99253610      +     MMS19          -                   exp_4
#> 27:  10          90471211      +      LIPK          +                   exp_5
#> 28:   1           8744319      -      RERE          -                   exp_2
#> 29:   1          89274341      +      PKN2          +                   exp_2
#> 30:   1           8744319      -      RERE          -                   exp_3
#> 31:   1          89274341      +      PKN2          +                   exp_3
#> 32:   2          33962502      + LINC01317          +                   exp_2
#> 33:   2          42521808      +      EML4          +                   exp_2
#> 34:   2          33962502      + LINC01317          +                   exp_3
#> 35:   5         152257562      + LINC01470          -                   exp_1
#> 36:   5          68342842      + LINC02198          -                   exp_1
#> 37:   9          77314074      +      RORB          +                   exp_5
#> 38:   8         137120365      +   KHDRBS3          +                   exp_2
#> 39:   8          87341169      -      WWP1          +                   exp_2
#> 40:   8          63731501      +    NKAIN3          +                   exp_3
#> 41:   8         137120365      +   KHDRBS3          +                   exp_5
#> 42:  12          88546429      -     TMTC3          +                   exp_2
#> 43:  12          88546429      -     TMTC3          +                   exp_5
#> 44:  18          69964312      -     CBLN2          -                   exp_4
#> 45:   X          41046822      +     USP9X          +                   exp_1
#> 46:   X          41046822      +     USP9X          +                   exp_4
#> 47:  22          17468972      -      GAB4          -                   exp_3
#> 48:  22          17468972      -      GAB4          -                   exp_5
#> 49:   3         137078019      -    IL20RB          +                   exp_1
#> 50:   3         157949340      -     RSRC1          +                   exp_2
#> 51:   3         137078019      -    IL20RB          +                   exp_2
#> 52:  16          53709727      -  RPGRIP1L          -                   exp_1
#> 53:  16          53709727      -  RPGRIP1L          -                   exp_5
#>     chr integration_locus strand  GeneName GeneStrand CompleteAmplificationID
#>         Value Value_RelAbundance Value_PercAbundance
#>  1: 94925.982        0.132693720          13.2693720
#>  2: 61143.203        0.098832448           9.8832448
#>  3: 68981.022        0.109473298          10.9473298
#>  4: 99324.680        0.270349560          27.0349560
#>  5: 83873.375        0.117243667          11.7243667
#>  6: 60410.179        0.095871318           9.5871318
#>  7: 47565.936        0.075487427           7.5487427
#>  8: 68463.189        0.110664541          11.0664541
#>  9: 46666.038        0.100913331          10.0913331
#> 10: 84163.039        0.136041926          13.6041926
#> 11: 70063.925        0.190705383          19.0705383
#> 12: 29285.895        0.046476892           4.6476892
#> 13: 25408.730        0.054945302           5.4945302
#> 14: 76745.114        0.165958059          16.5958059
#> 15: 42858.058        0.068015996           6.8015996
#> 16:  3736.397        0.008079801           0.8079801
#> 17: 29272.930        0.079677314           7.9677314
#> 18: 36743.994        0.051363148           5.1363148
#> 19: 34218.930        0.073996987           7.3996987
#> 20: 74588.496        0.120565546          12.0565546
#> 21: 63860.204        0.101346528          10.1346528
#> 22: 79029.953        0.170898928          17.0898928
#> 23:  8036.752        0.012990682           1.2990682
#> 24: 25255.111        0.035303239           3.5303239
#> 25: 97414.558        0.210654858          21.0654858
#> 26: 82334.313        0.178044466          17.8044466
#> 27: 30430.395        0.049187976           4.9187976
#> 28: 85050.007        0.118888440          11.8888440
#> 29: 12566.615        0.017566433           1.7566433
#> 30: 36910.659        0.058577438           5.8577438
#> 31: 61319.669        0.097314685           9.7314685
#> 32: 20246.969        0.028302532           2.8302532
#> 33: 73181.449        0.102297796          10.2297796
#> 34: 98047.224        0.155601535          15.5601535
#> 35: 92426.509        0.251573586          25.1573586
#> 36:  4388.195        0.011944127           1.1944127
#> 37: 55204.486        0.089233050           8.9233050
#> 38: 39071.423        0.054616580           5.4616580
#> 39: 67667.341        0.094589817           9.4589817
#> 40: 38778.586        0.061541849           6.1541849
#> 41: 92915.404        0.150189332          15.0189332
#> 42: 19758.601        0.027619859           2.7619859
#> 43: 39607.146        0.064021362           6.4021362
#> 44:  4799.085        0.010377819           1.0377819
#> 45: 37705.982        0.102631043          10.2631043
#> 46: 12083.681        0.026130449           2.6130449
#> 47: 82099.900        0.130293036          13.0293036
#> 48: 10109.592        0.016341239           1.6341239
#> 49: 23931.627        0.065138944           6.5138944
#> 50: 59266.776        0.082846960           8.2846960
#> 51: 97768.952        0.136667809          13.6667809
#> 52: 10279.687        0.027980043           2.7980043
#> 53: 93993.451        0.151931898          15.1931898
#>         Value Value_RelAbundance Value_PercAbundance

# Keeping totals as a separate data frame
abundance2 <- compute_abundance(matrix, keep_totals = "df")
abundance2

#> $abundance_df
#>     chr integration_locus strand  GeneName GeneStrand CompleteAmplificationID
#>  1:  20          39085641      +      MAFB          -                   exp_2
#>  2:  20          39085641      +      MAFB          -                   exp_5
#>  3:   7         155446768      +     RBM33          +                   exp_3
#>  4:  19           4049601      +    ZBTB7A          -                   exp_1
#>  5:  19           4049601      +    ZBTB7A          -                   exp_2
#>  6:  19          12278742      -    ZNF136          +                   exp_3
#>  7:  19           4049601      +    ZBTB7A          -                   exp_3
#>  8:  19          12278742      -    ZNF136          +                   exp_5
#>  9:   6           1865825      -      GMDS          -                   exp_4
#> 10:   6          70219927      -    ADGRB3          +                   exp_5
#> 11:  17          79831324      +   ARHGDIA          -                   exp_1
#> 12:  17           7113678      +      DLG4          -                   exp_3
#> 13:  17          56480574      +     RNF43          -                   exp_4
#> 14:  13         115089373      +    CHAMP1          +                   exp_4
#> 15:   4         127380734      -   MIR2054          +                   exp_3
#> 16:   4          46971398      -    GABRA4          -                   exp_4
#> 17:  11          57153892      +      PRG2          -                   exp_1
#> 18:  11          57153892      +      PRG2          -                   exp_2
#> 19:  11          57153892      +      PRG2          -                   exp_4
#> 20:  11          57153892      +      PRG2          -                   exp_5
#> 21:  15          56979473      -   ZNF280D          -                   exp_3
#> 22:  15          76209162      -    FBXO22          +                   exp_4
#> 23:  15          56979473      -   ZNF280D          -                   exp_5
#> 24:  10          99253610      +     MMS19          -                   exp_2
#> 25:  10          38314134      -    ZNF33A          +                   exp_4
#> 26:  10          99253610      +     MMS19          -                   exp_4
#> 27:  10          90471211      +      LIPK          +                   exp_5
#> 28:   1           8744319      -      RERE          -                   exp_2
#> 29:   1          89274341      +      PKN2          +                   exp_2
#> 30:   1           8744319      -      RERE          -                   exp_3
#> 31:   1          89274341      +      PKN2          +                   exp_3
#> 32:   2          33962502      + LINC01317          +                   exp_2
#> 33:   2          42521808      +      EML4          +                   exp_2
#> 34:   2          33962502      + LINC01317          +                   exp_3
#> 35:   5         152257562      + LINC01470          -                   exp_1
#> 36:   5          68342842      + LINC02198          -                   exp_1
#> 37:   9          77314074      +      RORB          +                   exp_5
#> 38:   8         137120365      +   KHDRBS3          +                   exp_2
#> 39:   8          87341169      -      WWP1          +                   exp_2
#> 40:   8          63731501      +    NKAIN3          +                   exp_3
#> 41:   8         137120365      +   KHDRBS3          +                   exp_5
#> 42:  12          88546429      -     TMTC3          +                   exp_2
#> 43:  12          88546429      -     TMTC3          +                   exp_5
#> 44:  18          69964312      -     CBLN2          -                   exp_4
#> 45:   X          41046822      +     USP9X          +                   exp_1
#> 46:   X          41046822      +     USP9X          +                   exp_4
#> 47:  22          17468972      -      GAB4          -                   exp_3
#> 48:  22          17468972      -      GAB4          -                   exp_5
#> 49:   3         137078019      -    IL20RB          +                   exp_1
#> 50:   3         157949340      -     RSRC1          +                   exp_2
#> 51:   3         137078019      -    IL20RB          +                   exp_2
#> 52:  16          53709727      -  RPGRIP1L          -                   exp_1
#> 53:  16          53709727      -  RPGRIP1L          -                   exp_5
#>     chr integration_locus strand  GeneName GeneStrand CompleteAmplificationID
#>         Value Value_RelAbundance Value_PercAbundance
#>  1: 94925.982        0.132693720          13.2693720
#>  2: 61143.203        0.098832448           9.8832448
#>  3: 68981.022        0.109473298          10.9473298
#>  4: 99324.680        0.270349560          27.0349560
#>  5: 83873.375        0.117243667          11.7243667
#>  6: 60410.179        0.095871318           9.5871318
#>  7: 47565.936        0.075487427           7.5487427
#>  8: 68463.189        0.110664541          11.0664541
#>  9: 46666.038        0.100913331          10.0913331
#> 10: 84163.039        0.136041926          13.6041926
#> 11: 70063.925        0.190705383          19.0705383
#> 12: 29285.895        0.046476892           4.6476892
#> 13: 25408.730        0.054945302           5.4945302
#> 14: 76745.114        0.165958059          16.5958059
#> 15: 42858.058        0.068015996           6.8015996
#> 16:  3736.397        0.008079801           0.8079801
#> 17: 29272.930        0.079677314           7.9677314
#> 18: 36743.994        0.051363148           5.1363148
#> 19: 34218.930        0.073996987           7.3996987
#> 20: 74588.496        0.120565546          12.0565546
#> 21: 63860.204        0.101346528          10.1346528
#> 22: 79029.953        0.170898928          17.0898928
#> 23:  8036.752        0.012990682           1.2990682
#> 24: 25255.111        0.035303239           3.5303239
#> 25: 97414.558        0.210654858          21.0654858
#> 26: 82334.313        0.178044466          17.8044466
#> 27: 30430.395        0.049187976           4.9187976
#> 28: 85050.007        0.118888440          11.8888440
#> 29: 12566.615        0.017566433           1.7566433
#> 30: 36910.659        0.058577438           5.8577438
#> 31: 61319.669        0.097314685           9.7314685
#> 32: 20246.969        0.028302532           2.8302532
#> 33: 73181.449        0.102297796          10.2297796
#> 34: 98047.224        0.155601535          15.5601535
#> 35: 92426.509        0.251573586          25.1573586
#> 36:  4388.195        0.011944127           1.1944127
#> 37: 55204.486        0.089233050           8.9233050
#> 38: 39071.423        0.054616580           5.4616580
#> 39: 67667.341        0.094589817           9.4589817
#> 40: 38778.586        0.061541849           6.1541849
#> 41: 92915.404        0.150189332          15.0189332
#> 42: 19758.601        0.027619859           2.7619859
#> 43: 39607.146        0.064021362           6.4021362
#> 44:  4799.085        0.010377819           1.0377819
#> 45: 37705.982        0.102631043          10.2631043
#> 46: 12083.681        0.026130449           2.6130449
#> 47: 82099.900        0.130293036          13.0293036
#> 48: 10109.592        0.016341239           1.6341239
#> 49: 23931.627        0.065138944           6.5138944
#> 50: 59266.776        0.082846960           8.2846960
#> 51: 97768.952        0.136667809          13.6667809
#> 52: 10279.687        0.027980043           2.7980043
#> 53: 93993.451        0.151931898          15.1931898
#>         Value Value_RelAbundance Value_PercAbundance
#> 
#> $quant_totals
#> # A tibble: 5 x 2
#>   CompleteAmplificationID Value_tot
#>   <fct>                       <dbl>
#> 1 exp_2                     715377.
#> 2 exp_5                     618655.
#> 3 exp_3                     630117.
#> 4 exp_1                     367394.
#> 5 exp_4                     462437.
#>