Integrations cumulative count in time by sample

This function computes the cumulative number of integrations observed in each sample at different time points by assuming that if an integration is observed at time point "t" then it is also observed in time point "t+1".

cumulative_count_union(
  x,
  association_file = NULL,
  timepoint_column = "TimePoint",
  key = c("SubjectID", "CellMarker", "Tissue", "TimePoint"),
  include_tp_zero = FALSE,
  zero = "0000",
  aggregate = FALSE,
  ...
)

Arguments

x
A simple integration matrix or an aggregated matrix (see details)

association_file
NULL or the association file for x if aggregate is set to TRUE

timepoint_column
What is the name of the time point column?

key
The aggregation key - must always contain the timepoint_column

include_tp_zero
Include timepoint 0?

zero
How is 0 coded in the data frame?

aggregate
Should x be aggregated?

...
Additional parameters to pass to aggregate_values_by_key

Value

A data frame

Details

Input data frame

The user can provide as input for the x parameter both a simple integration matrix AND setting the aggregate parameter to TRUE, or provide an already aggregated matrix via aggregate_values_by_key. If the user supplies a matrix to be aggregated the association_file parameter must not be NULL: aggregation will be done by an internal call to the aggregation function. If the user supplies an already aggregated matrix, the key parameter is the key used for aggregation - NOTE: for this operation is mandatory that the time point column is included in the key.

Assumptions on time point format

By using the functions provided by this package, when imported, an association file will be correctly formatted for future usage. In the formatting process there is also a padding operation performed on time points: this means the functions expects the time point column to be of type character and to be correctly padded with 0s. If the chosen column for time point is detected as numeric the function will attempt the conversion to character and automatic padding. If you choose to import the association file not using the import_association_file function, be sure to check the format of the chosen column to avoid undesired results.

See also

Examples

data("integration_matrices", package = "ISAnalytics")
data("association_file", package = "ISAnalytics")
aggreg <- aggregate_values_by_key(
    x = integration_matrices,
    association_file = association_file,
    value_cols = c("seqCount", "fragmentEstimate")
)
cumulative_count <- cumulative_count_union(aggreg)
cumulative_count

#> # A tibble: 20 × 5
#>    SubjectID CellMarker Tissue TimePoint count
#>    <chr>     <chr>      <chr>  <chr>     <int>
#>  1 PT001     MNC        BM     0030         54
#>  2 PT001     MNC        BM     0060        147
#>  3 PT001     MNC        BM     0090        179
#>  4 PT001     MNC        BM     0180        240
#>  5 PT001     MNC        BM     0360        240
#>  6 PT001     MNC        PB     0030         28
#>  7 PT001     MNC        PB     0060         77
#>  8 PT001     MNC        PB     0090        104
#>  9 PT001     MNC        PB     0180        121
#> 10 PT001     MNC        PB     0360        121
#> 11 PT002     MNC        BM     0030         98
#> 12 PT002     MNC        BM     0060        126
#> 13 PT002     MNC        BM     0090        141
#> 14 PT002     MNC        BM     0180        184
#> 15 PT002     MNC        BM     0360        265
#> 16 PT002     MNC        PB     0030         15
#> 17 PT002     MNC        PB     0060         26
#> 18 PT002     MNC        PB     0090         38
#> 19 PT002     MNC        PB     0180         62
#> 20 PT002     MNC        PB     0360        109