Expands integration matrix with the cumulative is union over time.

Given an input integration matrix that can be grouped over time, this function adds integrations in groups assuming that if an integration is observed at time point "t" then it is also observed in time point "t+1".

cumulative_is(
  x,
  key = c("SubjectID", "CellMarker", "Tissue", "TimePoint"),
  timepoint_col = "TimePoint",
  include_tp_zero = FALSE,
  keep_og_is = TRUE,
  expand = FALSE
)

Arguments

x

An integration matrix, ideally aggregated via aggregate_values_by_key()

key

The aggregation key used

timepoint_col

The name of the time point column

include_tp_zero

Should time point 0 be included?

keep_og_is

Keep original set of integrations as a separate column?

expand

If FALSE, for each group, the set of integration sites is returned in a separate column as a nested table, otherwise the resulting column is unnested.

Value

A data frame

See also

Examples

data("integration_matrices", package = "ISAnalytics")
data("association_file", package = "ISAnalytics")
aggreg <- aggregate_values_by_key(
    x = integration_matrices,
    association_file = association_file,
    value_cols = c("seqCount", "fragmentEstimate")
)
cumulated_is <- cumulative_is(aggreg)
cumulated_is
#>     SubjectID CellMarker Tissue TimePoint                  is
#>  1:     PT001        MNC     BM        30  <data.table[54x5]>
#>  2:     PT001        MNC     BM        60 <data.table[114x5]>
#>  3:     PT001        MNC     BM        90  <data.table[59x5]>
#>  4:     PT001        MNC     BM       180  <data.table[89x5]>
#>  5:     PT001        MNC     BM       360  <data.table[78x5]>
#>  6:     PT001        MNC     PB        30  <data.table[28x5]>
#>  7:     PT001        MNC     PB        60  <data.table[59x5]>
#>  8:     PT001        MNC     PB        90  <data.table[48x5]>
#>  9:     PT001        MNC     PB       180  <data.table[29x5]>
#> 10:     PT001        MNC     PB       360  <data.table[43x5]>
#> 11:     PT002        MNC     BM        30  <data.table[98x5]>
#> 12:     PT002        MNC     BM        60  <data.table[33x5]>
#> 13:     PT002        MNC     BM        90  <data.table[40x5]>
#> 14:     PT002        MNC     BM       180  <data.table[69x5]>
#> 15:     PT002        MNC     BM       360  <data.table[91x5]>
#> 16:     PT002        MNC     PB        30  <data.table[15x5]>
#> 17:     PT002        MNC     PB        60  <data.table[18x5]>
#> 18:     PT002        MNC     PB        90  <data.table[24x5]>
#> 19:     PT002        MNC     PB       180  <data.table[32x5]>
#> 20:     PT002        MNC     PB       360  <data.table[53x5]>
#>           cumulative_is
#>  1:  <data.table[54x5]>
#>  2: <data.table[147x5]>
#>  3: <data.table[179x5]>
#>  4: <data.table[240x5]>
#>  5: <data.table[240x5]>
#>  6:  <data.table[28x5]>
#>  7:  <data.table[77x5]>
#>  8: <data.table[104x5]>
#>  9: <data.table[121x5]>
#> 10: <data.table[121x5]>
#> 11:  <data.table[98x5]>
#> 12: <data.table[126x5]>
#> 13: <data.table[141x5]>
#> 14: <data.table[184x5]>
#> 15: <data.table[265x5]>
#> 16:  <data.table[15x5]>
#> 17:  <data.table[26x5]>
#> 18:  <data.table[38x5]>
#> 19:  <data.table[62x5]>
#> 20: <data.table[109x5]>