Default metadata aggregation function table

A default columns-function specifications for aggregate_metadata

default_meta_agg()

Value

A data frame

Details

This data frame contains four columns:

Column: holds the name of the column in the association file that should be processed

Function: contains either the name of a function (e.g. mean) or a purrr-style lambda (e.g. ~ mean(.x, na.rm = TRUE)). This function will be applied to the corresponding column specified in Column

Args: optional additional arguments to pass to the corresponding function. This is relevant ONLY if the corresponding Function is a simple function and not a purrr-style lambda.

Output_colname: a glue specification that will be used to determine a unique output column name. See glue for more details.

See also

Examples

default_meta_agg()
#> # A tibble: 12 × 4
#>    Column               Function  Args  Output_colname
#>    <chr>                <list>    <lgl> <chr>         
#>  1 FusionPrimerPCRDate  <formula> NA    {.col}_min    
#>  2 LinearPCRDate        <formula> NA    {.col}_min    
#>  3 VCN                  <formula> NA    {.col}_avg    
#>  4 ng DNA corrected     <formula> NA    {.col}_avg    
#>  5 Kapa                 <formula> NA    {.col}_avg    
#>  6 ng DNA corrected     <formula> NA    {.col}_sum    
#>  7 ulForPool            <formula> NA    {.col}_sum    
#>  8 BARCODE_MUX          <formula> NA    {.col}_sum    
#>  9 TRIMMING_FINAL_LTRLC <formula> NA    {.col}_sum    
#> 10 LV_MAPPED            <formula> NA    {.col}_sum    
#> 11 BWA_MAPPED_OVERALL   <formula> NA    {.col}_sum    
#> 12 ISS_MAPPED_PP        <formula> NA    {.col}_sum