Generate a default folder structure, following VISPA2 standards

The function produces a folder structure in the file system at the provided path that respects VISPA2 standards, with package-included data.

Usage

generate_default_folder_structure(
  type = "correct",
  dir = tempdir(),
  af = "default",
  matrices = "default"
)

Arguments

type

One value between "correct", "incorrect" and "both". Tells the function wheter to produce a correct structure or introduce some errors (mainly for testing purposes).

dir

Path to the folder in which the structure will be produced

af

Either "default" for the association file provided as example in the package or a custom association file as a data frame

matrices

Either "default" for integration matrices provided as example in the package or a custom multi-quantification matrix

Value

A named list containing the path to the association file and the path to the top level folder(s) of the structure

Required tags

The function will explicitly check for the presence of these tags:

project_id

tag_seq

vispa_concatenate

See also

Examples

fs_path <- generate_default_folder_structure(type = "correct")
fs_path
#> $af
#> /tmp/RtmpoihKVt/asso_file.tsv
#> 
#> $root
#> /tmp/RtmpoihKVt/fs
#>