Import a dynamic vars settings profile.

The function allows the import of an existing dynamic vars profile in json format. This is a quick and convenient way to set up the workflow, alternative to specifying lookup tables manually through the corresponding setter functions. For more details, refer to the dedicated vignette vignette("workflow_start", package="ISAnalytics").

Usage

import_ISA_settings(path)

Arguments

path

The path to the json file on disk

Value

NULL

See also

Examples

tmp_folder <- tempdir()
export_ISA_settings(tmp_folder, "DEFAULT")
#> Settings profile correctly saved
#> ℹ Saved at: /tmp/RtmpXJcUps/DEFAULT_ISAsettings.json
import_ISA_settings(fs::path(tmp_folder, "DEFAULT_ISAsettings.json"))
#> Mandatory IS vars successfully changed
#> Annotation IS vars successfully changed
#> Association file columns specs successfully changed
#> ISS stats specs successfully changed
#> Matrix suffixes specs successfully changed
reset_dyn_vars_config()
#> Mandatory IS vars reset to default
#> Annotation IS vars reset to default
#> Association file columns specs reset to default
#> ISS stats specs reset to default
#> Matrix suffixes specs reset to default