Import integration matrices from paths in the association file.

The function offers a convenient way of importing multiple integration matrices in an automated or semi-automated way. For more details see the "How to use import functions" vignette: vignette("import_functions_howto", package = "ISAnalytics")

Usage

import_parallel_Vispa2Matrices(
  association_file,
  quantification_type,
  matrix_type = "annotated",
  workers = 2,
  multi_quant_matrix = TRUE,
  report_path = default_report_path(),
  patterns = NULL,
  matching_opt = matching_options(),
  mode = c("AUTO", "INTERACTIVE"),
  ...
)

Arguments

association_file

Data frame imported via import_association_file (with file system alignment) or a string containing the path to the association file on disk.

quantification_type

A vector of requested quantification_types. Possible choices are quantification_types

matrix_type

A single string representing the type of matrices to be imported. Can only be one in "annotated" or "not_annotated".

workers

A single integer representing the number of parallel workers to use for the import

multi_quant_matrix

If set to TRUE will produce a multi-quantification matrix through comparison_matrix instead of a list.

report_path

The path where the report file should be saved. Can be a folder, a file or NULL if no report should be produced. Defaults to {user_home}/ISAnalytics_reports.

patterns

Relevant only if argument mode is set to AUTO. A character vector of additional patterns to match on file names. Please note that patterns must be regular expressions. Can be NULL if no patterns need to be matched.

matching_opt

Relevant only if argument mode is set to AUTO. A single value between matching_options

mode

A single value between AUTO and INTERACTIVE. If INTERACTIVE, the function will ask for input from the user on console, otherwise the process is fully automated (with limitations, see vignette).

...

<dynamic-dots> Additional named arguments to pass to ìmport_association_file and comparison_matrix

Value

Either a multi-quantification matrix or a list of integration matrices

See also

Examples

fs_path <- system.file("extdata", "fs.zip", package = "ISAnalytics")
fs <- unzip_file_system(fs_path, "fs")
af_path <- system.file("extdata", "asso.file.tsv.gz",
    package = "ISAnalytics"
)
af <- import_association_file(af_path,
    root = fs,
    import_iss = FALSE,
    report_path = NULL
)
matrices <- import_parallel_Vispa2Matrices(af,
    c("seqCount", "fragmentEstimate"),
    mode = "AUTO", report_path = NULL
)
head(matrices)
#> # A tibble: 6 × 8
#>   chr   integration_locus strand GeneName     GeneStrand CompleteAmplificationID
#>   <chr>             <int> <chr>  <chr>        <chr>      <fct>                  
#> 1 16             68164148 +      NFATC3       +          PJ01_POOL01_LTR75LC38_…
#> 2 4             129390130 +      LOC100507487 +          PJ01_POOL01_LTR75LC38_…
#> 3 5              84009671 -      EDIL3        -          PJ01_POOL01_LTR75LC38_…
#> 4 12             54635693 -      CBX5         -          PJ01_POOL01_LTR75LC38_…
#> 5 2             181930711 +      UBE2E3       +          PJ01_POOL01_LTR75LC38_…
#> 6 20             35920986 +      MANBAL       +          PJ01_POOL01_LTR75LC38_…
#> # … with 2 more variables: fragmentEstimate <dbl>, seqCount <int>