Find the source of IS by evaluating sharing.

The function computes the sharing between a reference group of interest for each time point and a selection of groups of interest. In this way it is possible to observe the percentage of shared integration sites between reference and each group and identify in which time point a certain IS was observed for the first time.

iss_source(
  reference,
  selection,
  ref_group_key = c("SubjectID", "CellMarker", "Tissue", "TimePoint"),
  selection_group_key = c("SubjectID", "CellMarker", "Tissue", "TimePoint"),
  timepoint_column = "TimePoint",
  by_subject = TRUE,
  subject_column = "SubjectID"
)

Arguments

reference
A data frame containing one or more groups of reference. Groups are identified by ref_group_key

selection
A data frame containing one or more groups of interest to compare. Groups are identified by selection_group_key

ref_group_key
Character vector of column names that identify a unique group in the reference data frame

selection_group_key
Character vector of column names that identify a unique group in the selection data frame

timepoint_column
Name of the column holding time point info?

by_subject
Should calculations be performed for each subject separately?

subject_column
Name of the column holding subjects information. Relevant only if by_subject = TRUE

Value

A list of data frames or a data frame

See also

Examples

data("integration_matrices", package = "ISAnalytics")
data("association_file", package = "ISAnalytics")
aggreg <- aggregate_values_by_key(
    x = rlang::current_env()$integration_matrices,
    association_file = rlang::current_env()$association_file,
    value_cols = c("seqCount", "fragmentEstimate")
)

#> Error in aggregate_values_by_key(x = rlang::current_env()$integration_matrices,     association_file = rlang::current_env()$association_file,     value_cols = c("seqCount", "fragmentEstimate")): is.data.frame(x) || is.list(x) is not TRUE

df1 <- aggreg %>%
    dplyr::filter(.data$Tissue == "BM")

#> Error in dplyr::filter(., .data$Tissue == "BM"): object 'aggreg' not found

df2 <- aggreg %>%
    dplyr::filter(.data$Tissue == "PB")

#> Error in dplyr::filter(., .data$Tissue == "PB"): object 'aggreg' not found

source <- iss_source(df1, df2)

#> Error in is.data.frame(reference): object 'df1' not found

source

#> function (file, local = FALSE, echo = verbose, print.eval = echo, 
#>     exprs, spaced = use_file, verbose = getOption("verbose"), 
#>     prompt.echo = getOption("prompt"), max.deparse.length = 150, 
#>     width.cutoff = 60L, deparseCtrl = "showAttributes", chdir = FALSE, 
#>     encoding = getOption("encoding"), continue.echo = getOption("continue"), 
#>     skip.echo = 0, keep.source = getOption("keep.source")) 
#> {
#>     envir <- if (isTRUE(local)) 
#>         parent.frame()
#>     else if (isFALSE(local)) 
#>         .GlobalEnv
#>     else if (is.environment(local)) 
#>         local
#>     else stop("'local' must be TRUE, FALSE or an environment")
#>     if (!missing(echo)) {
#>         if (!is.logical(echo)) 
#>             stop("'echo' must be logical")
#>         if (!echo && verbose) {
#>             warning("'verbose' is TRUE, 'echo' not; ... coercing 'echo <- TRUE'")
#>             echo <- TRUE
#>         }
#>     }
#>     if (verbose) {
#>         cat("'envir' chosen:")
#>         print(envir)
#>     }
#>     if (use_file <- missing(exprs)) {
#>         ofile <- file
#>         from_file <- FALSE
#>         srcfile <- NULL
#>         if (is.character(file)) {
#>             have_encoding <- !missing(encoding) && encoding != 
#>                 "unknown"
#>             if (identical(encoding, "unknown")) {
#>                 enc <- utils::localeToCharset()
#>                 encoding <- enc[length(enc)]
#>             }
#>             else enc <- encoding
#>             if (length(enc) > 1L) {
#>                 encoding <- NA
#>                 owarn <- options(warn = 2)
#>                 for (e in enc) {
#>                   if (is.na(e)) 
#>                     next
#>                   zz <- file(file, encoding = e)
#>                   res <- tryCatch(readLines(zz, warn = FALSE), 
#>                     error = identity)
#>                   close(zz)
#>                   if (!inherits(res, "error")) {
#>                     encoding <- e
#>                     break
#>                   }
#>                 }
#>                 options(owarn)
#>             }
#>             if (is.na(encoding)) 
#>                 stop("unable to find a plausible encoding")
#>             if (verbose) 
#>                 cat(gettextf("encoding = \"%s\" chosen", encoding), 
#>                   "\n", sep = "")
#>             if (file == "") {
#>                 file <- stdin()
#>                 srcfile <- "<stdin>"
#>             }
#>             else {
#>                 filename <- file
#>                 file <- file(filename, "r", encoding = encoding)
#>                 on.exit(close(file))
#>                 if (isTRUE(keep.source)) {
#>                   lines <- readLines(file, warn = FALSE)
#>                   on.exit()
#>                   close(file)
#>                   srcfile <- srcfilecopy(filename, lines, file.mtime(filename)[1], 
#>                     isFile = TRUE)
#>                 }
#>                 else {
#>                   from_file <- TRUE
#>                   srcfile <- filename
#>                 }
#>                 loc <- utils::localeToCharset()[1L]
#>                 encoding <- if (have_encoding) 
#>                   switch(loc, `UTF-8` = "UTF-8", `ISO8859-1` = "latin1", 
#>                     "unknown")
#>                 else "unknown"
#>             }
#>         }
#>         else {
#>             lines <- readLines(file, warn = FALSE)
#>             srcfile <- if (isTRUE(keep.source)) 
#>                 srcfilecopy(deparse1(substitute(file), ""), lines)
#>             else deparse1(substitute(file), "")
#>         }
#>         exprs <- if (!from_file) {
#>             if (length(lines)) 
#>                 .Internal(parse(stdin(), n = -1, lines, "?", 
#>                   srcfile, encoding))
#>             else expression()
#>         }
#>         else .Internal(parse(file, n = -1, NULL, "?", srcfile, 
#>             encoding))
#>         on.exit()
#>         if (from_file) 
#>             close(file)
#>         if (verbose) 
#>             cat("--> parsed", length(exprs), "expressions; now eval(.)ing them:\n")
#>         if (chdir) {
#>             if (is.character(ofile)) {
#>                 if (grepl("^(ftp|http|file)://", ofile)) 
#>                   warning("'chdir = TRUE' makes no sense for a URL")
#>                 else if ((path <- dirname(ofile)) != ".") {
#>                   owd <- getwd()
#>                   if (is.null(owd)) 
#>                     stop("cannot 'chdir' as current directory is unknown")
#>                   on.exit(setwd(owd), add = TRUE)
#>                   setwd(path)
#>                 }
#>             }
#>             else {
#>                 warning("'chdir = TRUE' makes no sense for a connection")
#>             }
#>         }
#>     }
#>     else {
#>         if (!missing(file)) 
#>             stop("specify either 'file' or 'exprs' but not both")
#>         if (!is.expression(exprs)) 
#>             exprs <- as.expression(exprs)
#>     }
#>     Ne <- length(exprs)
#>     if (echo) {
#>         sd <- "\""
#>         nos <- "[^\"]*"
#>         oddsd <- paste0("^", nos, sd, "(", nos, sd, nos, sd, 
#>             ")*", nos, "$")
#>         trySrcLines <- function(srcfile, showfrom, showto) {
#>             tryCatch(suppressWarnings(getSrcLines(srcfile, showfrom, 
#>                 showto)), error = function(e) character())
#>         }
#>     }
#>     yy <- NULL
#>     lastshown <- 0
#>     srcrefs <- attr(exprs, "srcref")
#>     if (verbose && !is.null(srcrefs)) {
#>         cat("has srcrefs:\n")
#>         utils::str(srcrefs)
#>     }
#>     for (i in seq_len(Ne + echo)) {
#>         tail <- i > Ne
#>         if (!tail) {
#>             if (verbose) 
#>                 cat("\n>>>> eval(expression_nr.", i, ")\n\t\t =================\n")
#>             ei <- exprs[i]
#>         }
#>         if (echo) {
#>             nd <- 0
#>             srcref <- if (tail) 
#>                 attr(exprs, "wholeSrcref")
#>             else if (i <= length(srcrefs)) 
#>                 srcrefs[[i]]
#>             if (!is.null(srcref)) {
#>                 if (i == 1) 
#>                   lastshown <- min(skip.echo, srcref[3L] - 1)
#>                 if (lastshown < srcref[3L]) {
#>                   srcfile <- attr(srcref, "srcfile")
#>                   dep <- trySrcLines(srcfile, lastshown + 1, 
#>                     srcref[3L])
#>                   if (length(dep)) {
#>                     leading <- if (tail) 
#>                       length(dep)
#>                     else srcref[1L] - lastshown
#>                     lastshown <- srcref[3L]
#>                     while (length(dep) && grepl("^[[:blank:]]*$", 
#>                       dep[1L])) {
#>                       dep <- dep[-1L]
#>                       leading <- leading - 1L
#>                     }
#>                     dep <- paste0(rep.int(c(prompt.echo, continue.echo), 
#>                       pmax(0L, c(leading, length(dep) - leading))), 
#>                       dep, collapse = "\n")
#>                     nd <- nchar(dep, "c")
#>                   }
#>                   else srcref <- NULL
#>                 }
#>             }
#>             if (is.null(srcref)) {
#>                 if (!tail) {
#>                   dep <- substr(paste(deparse(ei, width.cutoff = width.cutoff, 
#>                     control = deparseCtrl), collapse = "\n"), 
#>                     12L, 1000000L)
#>                   dep <- paste0(prompt.echo, gsub("\n", paste0("\n", 
#>                     continue.echo), dep, fixed = TRUE))
#>                   nd <- nchar(dep, "c") - 1L
#>                 }
#>             }
#>             if (nd) {
#>                 do.trunc <- nd > max.deparse.length
#>                 dep <- substr(dep, 1L, if (do.trunc) 
#>                   max.deparse.length
#>                 else nd)
#>                 cat(if (spaced) 
#>                   "\n", dep, if (do.trunc) 
#>                   paste(if (grepl(sd, dep) && grepl(oddsd, dep)) 
#>                     " ...\" ..."
#>                   else " ....", "[TRUNCATED] "), "\n", sep = "")
#>             }
#>         }
#>         if (!tail) {
#>             yy <- withVisible(eval(ei, envir))
#>             i.symbol <- mode(ei[[1L]]) == "name"
#>             if (!i.symbol) {
#>                 curr.fun <- ei[[1L]][[1L]]
#>                 if (verbose) {
#>                   cat("curr.fun:")
#>                   utils::str(curr.fun)
#>                 }
#>             }
#>             if (verbose >= 2) {
#>                 cat(".... mode(ei[[1L]])=", mode(ei[[1L]]), "; paste(curr.fun)=")
#>                 utils::str(paste(curr.fun))
#>             }
#>             if (print.eval && yy$visible) {
#>                 if (isS4(yy$value)) 
#>                   methods::show(yy$value)
#>                 else print(yy$value)
#>             }
#>             if (verbose) 
#>                 cat(" .. after ", sQuote(deparse(ei, control = unique(c(deparseCtrl, 
#>                   "useSource")))), "\n", sep = "")
#>         }
#>     }
#>     invisible(yy)
#> }
#> <bytecode: 0x7f99dd332b38>
#> <environment: namespace:base>

ggplot2::ggplot(source$PT001, ggplot2::aes(x = as.factor(g2_TimePoint),
                                           y = sharing_perc, fill = g1)) +
    ggplot2::geom_col() +
    ggplot2::labs(x = "Time point", y = "Shared IS % with MNC BM",
                  title = "Source of is MNC BM vs MNC PB")

#> Error in source$PT001: object of type 'closure' is not subsettable