left_join implementation for ISADataFrame.

left_join implementation for ISADataFrame.

# S3 method for ISADataFrame
left_join(
  x,
  y,
  by = NULL,
  copy = FALSE,
  suffix = c(".x", ".y"),
  ...,
  keep = FALSE
)

Arguments

x
A pair of data frames, data frame extensions (e.g. a tibble), or lazy data frames (e.g. from dbplyr or dtplyr). See Methods, below, for more details.

y
A pair of data frames, data frame extensions (e.g. a tibble), or lazy data frames (e.g. from dbplyr or dtplyr). See Methods, below, for more details.

by
A character vector of variables to join by.

If NULL, the default, *_join() will perform a natural join, using all variables in common across x and y. A message lists the variables so that you can check they're correct; suppress the message by supplying by explicitly.

To join by different variables on x and y, use a named vector. For example, by = c("a" = "b") will match x$a to y$b.

To join by multiple variables, use a vector with length > 1. For example, by = c("a", "b") will match x$a to y$a and x$b to y$b. Use a named vector to match different variables in x and y. For example, by = c("a" = "b", "c" = "d") will match x$a to y$b and x$c to y$d.

To perform a cross-join, generating all combinations of x and y, use by = character().

copy
If x and y are not from the same data source, and copy is TRUE, then y will be copied into the same src as x. This allows you to join tables across srcs, but it is a potentially expensive operation so you must opt into it.

suffix
If there are non-joined duplicate variables in x and y, these suffixes will be added to the output to disambiguate them. Should be a character vector of length 2.

...
Other parameters passed onto methods.

keep
Should the join keys from both x and y be preserved in the output? Only applies to nest_join(), left_join(), right_join(), and full_join().

Value

An ISADataFrame

Examples

path <- system.file("extdata", "ex_annotated_ISMatrix.tsv.xz",
    package = "ISAnalytics"
)
isadf1 <- import_single_Vispa2Matrix(path)
isadf2 <- import_single_Vispa2Matrix(path)
isadf1 %>% dplyr::left_join(isadf2)

#> Joining, by = c("chr", "integration_locus", "strand", "GeneName", "GeneStrand", "CompleteAmplificationID", "Value")

#> mandatoryVars:  chr,  integration_locus,  strand 
#> metadata:  GeneName,  GeneStrand,  CompleteAmplificationID 
#> # A tibble: 53 x 7
#>    chr   integration_locus strand GeneName GeneStrand CompleteAmplificat…  Value
#>    <chr>             <int> <chr>  <chr>    <chr>      <chr>                <dbl>
#>  1 1               8744319 -      RERE     -          exp_2               85050.
#>  2 1               8744319 -      RERE     -          exp_3               36911.
#>  3 1              89274341 +      PKN2     +          exp_2               12567.
#>  4 1              89274341 +      PKN2     +          exp_3               61320.
#>  5 2              33962502 +      LINC013… +          exp_2               20247.
#>  6 2              33962502 +      LINC013… +          exp_3               98047.
#>  7 2              42521808 +      EML4     +          exp_2               73181.
#>  8 3             157949340 -      RSRC1    +          exp_2               59267.
#>  9 3             137078019 -      IL20RB   +          exp_1               23932.
#> 10 3             137078019 -      IL20RB   +          exp_2               97769.
#> # … with 43 more rows