Possible choices for the matching_opt parameter.

These are all the possible values for the matching_opt parameter in import_parallel_vispa2Matrices_auto.

Usage

matching_options()

Value

A vector of characters for matching_opt

Details

The values "ANY", "ALL" and "OPTIONAL", represent how the patterns should be matched, more specifically

ANY = look only for files that match AT LEAST one of the patterns specified

ALL = look only for files that match ALL of the patterns specified

OPTIONAL = look preferentially for files that match, in order, all patterns or any pattern and if no match is found return what is found (keep in mind that duplicates are discarded in automatic mode)

See also

Examples

opts <- matching_options()