Easily retrieve the name of the pcr id column.

The function is a shortcut to retrieve the currently set pcr id column name from the association file column tags look-up table. This column is needed every time a joining operation with metadata needs to be performed

Usage

pcr_id_column()

Value

The name of the column

See also

Examples

pcr_id_column()
#> [1] "CompleteAmplificationID"