Re-aligns matrices of other quantification types based on the processed sequence count matrix.

This function should be used to keep data consistent among the same analysis: if for some reason you removed the collisions by passing only the sequence count matrix to the remove_collisions function, you should call this function afterwards, providing a list of other quantification matrices. NOTE: if you provided a list of several quantification types to remove_collisions before, there is no need to call this function.

realign_after_collisions(sc_matrix, other_matrices)

Arguments

sc_matrix
The sequence count matrix already processed for collisions via remove_collisions

other_matrices
A named list of matrices to re-align. Names in the list must be quantification types (quantification_types()) except "seqCount".

Value

A named list with re-aligned matrices

Details

For more details on how to use collision removal functionality: vignette("Collision removal functionality", package = "ISAnalytics")

See also

Examples

op <- options("ISAnalytics.widgets" = FALSE)
path <- system.file("extdata", "ex_association_file.tsv",
    package = "ISAnalytics"
)
root_pth <- system.file("extdata", "fs.zip", package = "ISAnalytics")
root <- unzip_file_system(root_pth, "fs")
association_file <- import_association_file(path, root)
matrices <- import_parallel_Vispa2Matrices_auto(
    association_file, NULL,
    c("fragmentEstimate", "seqCount"), "annotated", 2, NULL, "ANY"
)
sc_matrix <- remove_collisions(matrices$seqCount, association_file)

#> Found additional data relative to some projects that are not included in the imported matrices. Here is a summary

#> # A tibble: 1 x 3
#>   ProjectID   PoolID       CompleteAmplificationID                              
#>   <chr>       <chr>        <chr>                                                
#> 1 PROJECT1101 ABY-LR-PL4-… PROJECT1101_ABY-LR-PL4-POOL54_LTR51LC51_VA2020-mix14…

#> Identifying collisions...

#> Processing collisions...

#> Finished! You provided a single sequence count as matrix, to re-align other related matrices see ?realign_after_collisions

others <- matrices[!names(matrices) %in% "seqCount"]
aligned_matrices <- realign_after_collisions(sc_matrix, others)
options(op)