Computes user specified functions on numerical columns and updates the metadata data frame accordingly.

The function operates on a data frame by grouping the content by the sample key and computing every function specified on every column in the value_columns parameter. After that the metadata data frame is updated by including the computed results as columns for the corresponding key. For this reason it's required that both x and metadata have the same sample key, and it's particularly important if the user is working with previously aggregated data. For example:

### Importing association file and matrices
path_AF <- system.file("extdata", "ex_association_file.tsv",
package = "ISAnalytics")
root_correct <- system.file("extdata", "fs.zip",
package = "ISAnalytics")
root_correct <- unzip_file_system(root_correct, "fs")

association_file <- import_association_file(path_AF, root_correct)
matrices <- import_parallel_Vispa2Matrices_auto(
association_file = association_file , root = NULL,
quantification_type = c("seqCount","fragmentEstimate"),
matrix_type = "annotated", workers = 2, patterns = NULL,
matching_opt = "ANY", dates_format = "dmy")

### Aggregating data (both by same key)
aggreggated_x <- aggregate_values_by_key(matrices$seqCount,
association_file)
aggregated_meta <- aggregate_metadata(association_file)

### Sample statistics
sample_stats <- sample_statistics(x = aggregated_x,
metadata = aggregated_meta,
sample_key = c("SubjectID", "CellMarker","Tissue", "TimePoint"))

sample_statistics(
  x,
  metadata,
  sample_key = "CompleteAmplificationID",
  value_columns = "Value",
  functions = default_stats()
)

Arguments

x
A data frame

metadata
The metadata data frame

sample_key
Character vector representing the key for identifying a sample

value_columns
THe name of the columns to be computed, must be numeric or integer

functions
A named list of function or purrr-style lambdas

Value

A list with modified x and metadata data frames

See also

Examples

op <- options(ISAnalytics.widgets = FALSE)

path_AF <- system.file("extdata", "ex_association_file.tsv",
    package = "ISAnalytics"
)
root_correct <- system.file("extdata", "fs.zip",
    package = "ISAnalytics"
)
root_correct <- unzip_file_system(root_correct, "fs")

association_file <- import_association_file(path_AF, root_correct,
    dates_format = "dmy"
)

#> *** Association file import summary ***
#> ℹ For detailed report please set option 'ISAnalytics.widgets' to TRUE
#> * Parsing problems detected: TRUE
#> * Date parsing problems: TRUE
#> * Column problems detected: TRUE
#> * NAs found in important columns: TRUE
#> * File system alignment: no problems detected

matrices <- import_parallel_Vispa2Matrices_auto(
    association_file = association_file, root = NULL,
    quantification_type = c("seqCount", "fragmentEstimate"),
    matrix_type = "annotated", workers = 2, patterns = NULL,
    matching_opt = "ANY", multi_quant_matrix = FALSE
)

#> [1] "--- REPORT: FILES IMPORTED ---"
#> # A tibble: 8 x 5
#>   ProjectID   concatenatePoolIDSeqRun Quantification_type
#>   <chr>       <chr>                   <chr>              
#> 1 CLOEXP      POOL6-1                 seqCount           
#> 2 PROJECT1100 ABX-LR-PL5-POOL14-1     seqCount           
#> 3 PROJECT1100 ABX-LR-PL6-POOL15-1     seqCount           
#> 4 PROJECT1101 ABY-LR-PL4-POOL54-2     seqCount           
#> 5 CLOEXP      POOL6-1                 fragmentEstimate   
#> 6 PROJECT1100 ABX-LR-PL5-POOL14-1     fragmentEstimate   
#> 7 PROJECT1100 ABX-LR-PL6-POOL15-1     fragmentEstimate   
#> 8 PROJECT1101 ABY-LR-PL4-POOL54-2     fragmentEstimate   
#>   Files_chosen                                                                  
#>   <fs::path>                                                                    
#> 1 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/VA/CLOEXP/quan…
#> 2 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/PROJECT1100/qu…
#> 3 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/PROJECT1100/qu…
#> 4 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/PROJECT1101/qu…
#> 5 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/VA/CLOEXP/quan…
#> 6 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/PROJECT1100/qu…
#> 7 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/PROJECT1100/qu…
#> 8 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/PROJECT1101/qu…
#>   Imported
#>   <lgl>   
#> 1 TRUE    
#> 2 TRUE    
#> 3 TRUE    
#> 4 TRUE    
#> 5 TRUE    
#> 6 TRUE    
#> 7 TRUE    
#> 8 TRUE    
#> [1] "--- SUMMARY OF FILES CHOSEN FOR IMPORT ---"
#> # A tibble: 8 x 4
#>   ProjectID   concatenatePoolIDSeqRun Quantification_type
#>   <chr>       <chr>                   <chr>              
#> 1 CLOEXP      POOL6-1                 seqCount           
#> 2 CLOEXP      POOL6-1                 fragmentEstimate   
#> 3 PROJECT1100 ABX-LR-PL5-POOL14-1     seqCount           
#> 4 PROJECT1100 ABX-LR-PL5-POOL14-1     fragmentEstimate   
#> 5 PROJECT1100 ABX-LR-PL6-POOL15-1     seqCount           
#> 6 PROJECT1100 ABX-LR-PL6-POOL15-1     fragmentEstimate   
#> 7 PROJECT1101 ABY-LR-PL4-POOL54-2     seqCount           
#> 8 PROJECT1101 ABY-LR-PL4-POOL54-2     fragmentEstimate   
#>   Files_chosen                                                                  
#>   <fs::path>                                                                    
#> 1 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/VA/CLOEXP/quan…
#> 2 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/VA/CLOEXP/quan…
#> 3 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/PROJECT1100/qu…
#> 4 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/PROJECT1100/qu…
#> 5 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/PROJECT1100/qu…
#> 6 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/PROJECT1100/qu…
#> 7 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/PROJECT1101/qu…
#> 8 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/PROJECT1101/qu…
#> [1] "--- INTEGRATION MATRICES FOUND REPORT ---"
#> # A tibble: 8 x 5
#>   ProjectID   concatenatePoolIDSeqRun Anomalies Quantification_type
#>   <chr>       <chr>                   <lgl>     <chr>              
#> 1 CLOEXP      POOL6-1                 FALSE     seqCount           
#> 2 CLOEXP      POOL6-1                 FALSE     fragmentEstimate   
#> 3 PROJECT1100 ABX-LR-PL5-POOL14-1     FALSE     seqCount           
#> 4 PROJECT1100 ABX-LR-PL5-POOL14-1     FALSE     fragmentEstimate   
#> 5 PROJECT1100 ABX-LR-PL6-POOL15-1     FALSE     seqCount           
#> 6 PROJECT1100 ABX-LR-PL6-POOL15-1     FALSE     fragmentEstimate   
#> 7 PROJECT1101 ABY-LR-PL4-POOL54-2     FALSE     seqCount           
#> 8 PROJECT1101 ABY-LR-PL4-POOL54-2     FALSE     fragmentEstimate   
#>   Files_found                                                                   
#>   <fs::path>                                                                    
#> 1 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/VA/CLOEXP/quan…
#> 2 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/VA/CLOEXP/quan…
#> 3 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/PROJECT1100/qu…
#> 4 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/PROJECT1100/qu…
#> 5 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/PROJECT1100/qu…
#> 6 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/PROJECT1100/qu…
#> 7 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/PROJECT1101/qu…
#> 8 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/PROJECT1101/qu…

stats <- sample_statistics(matrices$seqCount, association_file)
options(op)