Separate a multiple-quantification matrix into single quantification matrices.

The function separates a single multi-quantification integration matrix, obtained via comparison_matrix, into single quantification matrices as a named list of tibbles.

separate_quant_matrices(
  x,
  fragmentEstimate = "fragmentEstimate",
  seqCount = "seqCount",
  barcodeCount = "barcodeCount",
  cellCount = "cellCount",
  ShsCount = "ShsCount"
)

Arguments

x
Single integration matrix with multiple quantification value columns, likely obtained via comparison_matrix.

fragmentEstimate
Name of the fragment estimate values column in input

seqCount
Name of the sequence count values column in input

barcodeCount
Name of the barcode count values column in input

cellCount
Name of the cell count values column in input

ShsCount
Name of the shs count values column in input

Value

A named list of tibbles, where names are quantification types

See also

Examples

op <- options("ISAnalytics.widgets" = FALSE)
path <- system.file("extdata", "ex_association_file.tsv",
    package = "ISAnalytics"
)
root_pth <- system.file("extdata", "fs.zip", package = "ISAnalytics")
root <- unzip_file_system(root_pth, "fs")
matrices <- import_parallel_Vispa2Matrices_auto(
    path, root,
    c("fragmentEstimate", "seqCount"), "annotated", 2, NULL, "ANY",
    dates_format = "dmy"
)

#> [1] "--- REPORT: FILES IMPORTED ---"
#> # A tibble: 8 x 5
#>   ProjectID   concatenatePoolIDSeqRun Quantification_type
#>   <chr>       <chr>                   <chr>              
#> 1 CLOEXP      POOL6-1                 fragmentEstimate   
#> 2 PROJECT1100 ABX-LR-PL5-POOL14-1     fragmentEstimate   
#> 3 PROJECT1100 ABX-LR-PL6-POOL15-1     fragmentEstimate   
#> 4 PROJECT1101 ABY-LR-PL4-POOL54-2     fragmentEstimate   
#> 5 CLOEXP      POOL6-1                 seqCount           
#> 6 PROJECT1100 ABX-LR-PL5-POOL14-1     seqCount           
#> 7 PROJECT1100 ABX-LR-PL6-POOL15-1     seqCount           
#> 8 PROJECT1101 ABY-LR-PL4-POOL54-2     seqCount           
#>   Files_chosen                                                                  
#>   <fs::path>                                                                    
#> 1 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmphbUjQ7/fs/VA/CLOEXP/quan…
#> 2 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmphbUjQ7/fs/PROJECT1100/qu…
#> 3 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmphbUjQ7/fs/PROJECT1100/qu…
#> 4 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmphbUjQ7/fs/PROJECT1101/qu…
#> 5 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmphbUjQ7/fs/VA/CLOEXP/quan…
#> 6 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmphbUjQ7/fs/PROJECT1100/qu…
#> 7 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmphbUjQ7/fs/PROJECT1100/qu…
#> 8 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmphbUjQ7/fs/PROJECT1101/qu…
#>   Imported
#>   <lgl>   
#> 1 TRUE    
#> 2 TRUE    
#> 3 TRUE    
#> 4 TRUE    
#> 5 TRUE    
#> 6 TRUE    
#> 7 TRUE    
#> 8 TRUE    
#> [1] "--- SUMMARY OF FILES CHOSEN FOR IMPORT ---"
#> # A tibble: 8 x 4
#>   ProjectID   concatenatePoolIDSeqRun Quantification_type
#>   <chr>       <chr>                   <chr>              
#> 1 CLOEXP      POOL6-1                 fragmentEstimate   
#> 2 CLOEXP      POOL6-1                 seqCount           
#> 3 PROJECT1100 ABX-LR-PL5-POOL14-1     fragmentEstimate   
#> 4 PROJECT1100 ABX-LR-PL5-POOL14-1     seqCount           
#> 5 PROJECT1100 ABX-LR-PL6-POOL15-1     fragmentEstimate   
#> 6 PROJECT1100 ABX-LR-PL6-POOL15-1     seqCount           
#> 7 PROJECT1101 ABY-LR-PL4-POOL54-2     fragmentEstimate   
#> 8 PROJECT1101 ABY-LR-PL4-POOL54-2     seqCount           
#>   Files_chosen                                                                  
#>   <fs::path>                                                                    
#> 1 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmphbUjQ7/fs/VA/CLOEXP/quan…
#> 2 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmphbUjQ7/fs/VA/CLOEXP/quan…
#> 3 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmphbUjQ7/fs/PROJECT1100/qu…
#> 4 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmphbUjQ7/fs/PROJECT1100/qu…
#> 5 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmphbUjQ7/fs/PROJECT1100/qu…
#> 6 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmphbUjQ7/fs/PROJECT1100/qu…
#> 7 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmphbUjQ7/fs/PROJECT1101/qu…
#> 8 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmphbUjQ7/fs/PROJECT1101/qu…
#> [1] "--- INTEGRATION MATRICES FOUND REPORT ---"
#> # A tibble: 8 x 5
#>   ProjectID   concatenatePoolIDSeqRun Anomalies Quantification_type
#>   <chr>       <chr>                   <lgl>     <chr>              
#> 1 CLOEXP      POOL6-1                 FALSE     fragmentEstimate   
#> 2 CLOEXP      POOL6-1                 FALSE     seqCount           
#> 3 PROJECT1100 ABX-LR-PL5-POOL14-1     FALSE     fragmentEstimate   
#> 4 PROJECT1100 ABX-LR-PL5-POOL14-1     FALSE     seqCount           
#> 5 PROJECT1100 ABX-LR-PL6-POOL15-1     FALSE     fragmentEstimate   
#> 6 PROJECT1100 ABX-LR-PL6-POOL15-1     FALSE     seqCount           
#> 7 PROJECT1101 ABY-LR-PL4-POOL54-2     FALSE     fragmentEstimate   
#> 8 PROJECT1101 ABY-LR-PL4-POOL54-2     FALSE     seqCount           
#>   Files_found                                                                   
#>   <fs::path>                                                                    
#> 1 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmphbUjQ7/fs/VA/CLOEXP/quan…
#> 2 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmphbUjQ7/fs/VA/CLOEXP/quan…
#> 3 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmphbUjQ7/fs/PROJECT1100/qu…
#> 4 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmphbUjQ7/fs/PROJECT1100/qu…
#> 5 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmphbUjQ7/fs/PROJECT1100/qu…
#> 6 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmphbUjQ7/fs/PROJECT1100/qu…
#> 7 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmphbUjQ7/fs/PROJECT1101/qu…
#> 8 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmphbUjQ7/fs/PROJECT1101/qu…
#> [1] "--- ALIGNMENT RESULTS ---"
#> # A tibble: 4 x 4
#>   ProjectID   concatenatePoolIDSeqRun Found
#>   <chr>       <chr>                   <lgl>
#> 1 CLOEXP      POOL6-1                 TRUE 
#> 2 PROJECT1100 ABX-LR-PL5-POOL14-1     TRUE 
#> 3 PROJECT1100 ABX-LR-PL6-POOL15-1     TRUE 
#> 4 PROJECT1101 ABY-LR-PL4-POOL54-2     TRUE 
#>   Path                                                                          
#>   <chr>                                                                         
#> 1 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmphbUjQ7/fs/VA/CLOEXP/quan…
#> 2 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmphbUjQ7/fs/PROJECT1100/qu…
#> 3 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmphbUjQ7/fs/PROJECT1100/qu…
#> 4 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmphbUjQ7/fs/PROJECT1101/qu…

total_matrix <- comparison_matrix(matrices)
separated_matrix <- separate_quant_matrices(total_matrix)
options(op)