A utility function to unzip and use example file systems included in the package

This utility function is a simple shortcut to create a temporary directory, unzip and reference the examples file systems included in the package for testing purposes.

unzip_file_system(zipfile, name)

Arguments

zipfile
The zipped file to decompress

name
The name of the folder in the zipped archive ("fs" or "fserr")

Value

A path to reference

See also

Examples

root_pth <- system.file("extdata", "fs.zip", package = "ISAnalytics")
root <- unzip_file_system(root_pth, "fs")