Validator for ISADataFrame objects.

This is a devel function and should not be used interactively. The validator takes an ISADataFrame as input to check if the object was built correctly. More specifically:

Checks if the data frame contains the mandatory variables specified by the mandatoryVars attribute

Checks if the data frame contains the metadata variables specified by the metadata attribute (if not, generates only a warning)

Checks if there is at least one experimental data column: a column is considered experimental data if it's name is not contained both in mandatoryVars and metadata and the column contains numeric values.

If all checks pass the function returns TRUE, otherwise some kind of error is shown.

validate_ISADataFrame(x)

Arguments

x
the ISADataFrame object to validate

Value

'TRUE' if all checks pass, error otherwise

Examples

isadf <- new_ISADataFrame(list(
    chr = c(as.character(1:10)),
    integration_locus = runif(10, min = 100, max = 10000),
    strand = sample(c("+", "-"), 10, replace = TRUE),
    exp_1 = runif(10, min = 0, max = 10000),
    exp_2 = runif(10, min = 0, max = 10000),
    exp_3 = runif(10, min = 0, max = 10000)
))

validate_ISADataFrame(isadf)

#> [1] TRUE