Grubbs test for Common Insertion Sites (CIS).

Statistical approach for the validation of common insertion sites significance based on the comparison of the integration frequency at the CIS gene with respect to other genes contained in the surrounding genomic regions. For more details please refer to this paper: https://ashpublications.org/blood/article/117/20/5332/21206/Lentiviral-vector-common-integration-sites-in

Usage

CIS_grubbs(
  x,
  genomic_annotation_file = "hg19",
  grubbs_flanking_gene_bp = 1e+05,
  threshold_alpha = 0.05,
  by = NULL
)

Arguments

x

An integration matrix, must include the mandatory_IS_vars() columns and the annotation_IS_vars() columns

genomic_annotation_file

Database file for gene annotation, see details.

grubbs_flanking_gene_bp

Number of base pairs flanking a gene

threshold_alpha

Significance threshold

by

Either NULL or a character vector of column names. If not NULL, the function will perform calculations for each group and return a list of data frames with the results. E.g. for by = "SubjectID", CIS will be computed for each distinct SubjectID found in the table (of course, "SubjectID" column must be included in the input data frame).

Value

A data frame

Details

Genomic annotation file

This file is a data base, or more simply a .tsv file to import, with genes annotation for the specific genome. The annotations for the human genome (hg19) and murine genome (mm9) are already included in this package: to use one of them just set the argument genomic_annotation_file to either "hg19" or "mm9". If for any reason the user is performing an analysis on another genome, this file needs to be changed respecting the USCS Genome Browser format, meaning the input file headers should include:

name2 chrom strand min_txStart max_txEnd minmax_TxLen average_TxLen name min_cdsStart max_cdsEnd minmax_CdsLen average_CdsLen

See also

Examples

data("integration_matrices", package = "ISAnalytics")
cis <- CIS_grubbs(integration_matrices)
head(cis)
#> # A tibble: 6 × 40
#>   GeneName GeneStrand chr   n_IS_perGene min_bp_integration… max_bp_integration…
#>   <chr>    <chr>      <chr>        <int>               <dbl>               <dbl>
#> 1 ABHD16A  -          6                1            31654876            31654876
#> 2 ABHD2    +          15               1            89642998            89642998
#> 3 ACAP2    -          3                1           195107772           195107772
#> 4 ACOX1    -          17               3            73953431            73972782
#> 5 ACSM3    +          16               1            20777188            20777188
#> 6 ADD1     +          4                3             2859237             2871962
#> # … with 34 more variables: IS_span_bp <dbl>, avg_bp_integration_locus <dbl>,
#> #   median_bp_integration_locus <dbl>, distinct_orientations <int>, vars <dbl>,
#> #   n <dbl>, mean <dbl>, sd <dbl>, median <dbl>, trimmed <dbl>, mad <dbl>,
#> #   min <dbl>, max <dbl>, range <dbl>, skew <dbl>, kurtosis <dbl>, se <dbl>,
#> #   average_TxLen <dbl>, geneIS_frequency_byHitIS <dbl>,
#> #   raw_gene_integration_frequency <dbl>,
#> #   integration_frequency_withtolerance <dbl>, …