Grubbs test for Common Insertion Sites (CIS).

Statistical approach for the validation of common insertion sites significance based on the comparison of the integration frequency at the CIS gene with respect to other genes contained in the surrounding genomic regions. For more details please refer to this paper: https://ashpublications.org/blood/article/117/20/5332/21206/Lentiviral-vector-common-integration-sites-in

CIS_grubbs(
  x,
  genomic_annotation_file = system.file("extdata", "hg19.refGene.oracle.tsv.xz",
    package = "ISAnalytics"),
  grubbs_flanking_gene_bp = 1e+05,
  threshold_alpha = 0.05,
  add_standard_padjust = TRUE
)

Arguments

x
An integration matrix, must include the mandatory_IS_vars() columns and the annotation_IS_vars() columns

genomic_annotation_file
Database file for gene annotation, see details

grubbs_flanking_gene_bp
Number of base pairs flanking a gene

threshold_alpha
Significance threshold

add_standard_padjust
Compute the standard padjust?

Value

A data frame

Details

Genomic annotation file

This file is a data base, or more simply a .tsv file to import, with genes annotation for the specific genome. The annotations for the human genome (hg19) is already included in this package. If for any reason the user is performing an analysis on another genome, this file needs to be changed respecting the USCS Genome Browser format, meaning the input file headers should be:

## name2, chrom, strand 
##  min_txStart, max_txEnd, minmax_TxLen 
##  average_TxLen, name, min_cdsStart 
##  max_cdsEnd, minmax_CdsLen, average_CdsLen

See also

Examples

op <- options(ISAnalytics.widgets = FALSE)

path_AF <- system.file("extdata", "ex_association_file.tsv",
    package = "ISAnalytics"
)
root_correct <- system.file("extdata", "fs.zip",
    package = "ISAnalytics"
)
root_correct <- unzip_file_system(root_correct, "fs")

matrices <- import_parallel_Vispa2Matrices_auto(
    association_file = path_AF, root = root_correct,
    quantification_type = c("seqCount", "fragmentEstimate"),
    matrix_type = "annotated", workers = 2, patterns = NULL,
    matching_opt = "ANY",
    dates_format = "dmy"
)

#> *** Association file import summary ***
#> ℹ For detailed report please set option 'ISAnalytics.widgets' to TRUE
#> * Parsing problems detected: TRUE
#> * Date parsing problems: TRUE
#> * Column problems detected: TRUE
#> * NAs found in important columns: TRUE
#> * File system alignment: no problems detected

#> [1] "--- REPORT: FILES IMPORTED ---"
#> # A tibble: 8 x 5
#>   ProjectID   concatenatePoolIDSeqRun Quantification_type
#>   <chr>       <chr>                   <chr>              
#> 1 CLOEXP      POOL6-1                 seqCount           
#> 2 PROJECT1100 ABX-LR-PL5-POOL14-1     seqCount           
#> 3 PROJECT1100 ABX-LR-PL6-POOL15-1     seqCount           
#> 4 PROJECT1101 ABY-LR-PL4-POOL54-2     seqCount           
#> 5 CLOEXP      POOL6-1                 fragmentEstimate   
#> 6 PROJECT1100 ABX-LR-PL5-POOL14-1     fragmentEstimate   
#> 7 PROJECT1100 ABX-LR-PL6-POOL15-1     fragmentEstimate   
#> 8 PROJECT1101 ABY-LR-PL4-POOL54-2     fragmentEstimate   
#>   Files_chosen                                                                  
#>   <fs::path>                                                                    
#> 1 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/VA/CLOEXP/quan…
#> 2 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/PROJECT1100/qu…
#> 3 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/PROJECT1100/qu…
#> 4 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/PROJECT1101/qu…
#> 5 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/VA/CLOEXP/quan…
#> 6 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/PROJECT1100/qu…
#> 7 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/PROJECT1100/qu…
#> 8 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/PROJECT1101/qu…
#>   Imported
#>   <lgl>   
#> 1 TRUE    
#> 2 TRUE    
#> 3 TRUE    
#> 4 TRUE    
#> 5 TRUE    
#> 6 TRUE    
#> 7 TRUE    
#> 8 TRUE    
#> [1] "--- SUMMARY OF FILES CHOSEN FOR IMPORT ---"
#> # A tibble: 8 x 4
#>   ProjectID   concatenatePoolIDSeqRun Quantification_type
#>   <chr>       <chr>                   <chr>              
#> 1 CLOEXP      POOL6-1                 seqCount           
#> 2 CLOEXP      POOL6-1                 fragmentEstimate   
#> 3 PROJECT1100 ABX-LR-PL5-POOL14-1     seqCount           
#> 4 PROJECT1100 ABX-LR-PL5-POOL14-1     fragmentEstimate   
#> 5 PROJECT1100 ABX-LR-PL6-POOL15-1     seqCount           
#> 6 PROJECT1100 ABX-LR-PL6-POOL15-1     fragmentEstimate   
#> 7 PROJECT1101 ABY-LR-PL4-POOL54-2     seqCount           
#> 8 PROJECT1101 ABY-LR-PL4-POOL54-2     fragmentEstimate   
#>   Files_chosen                                                                  
#>   <fs::path>                                                                    
#> 1 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/VA/CLOEXP/quan…
#> 2 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/VA/CLOEXP/quan…
#> 3 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/PROJECT1100/qu…
#> 4 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/PROJECT1100/qu…
#> 5 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/PROJECT1100/qu…
#> 6 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/PROJECT1100/qu…
#> 7 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/PROJECT1101/qu…
#> 8 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/PROJECT1101/qu…
#> [1] "--- INTEGRATION MATRICES FOUND REPORT ---"
#> # A tibble: 8 x 5
#>   ProjectID   concatenatePoolIDSeqRun Anomalies Quantification_type
#>   <chr>       <chr>                   <lgl>     <chr>              
#> 1 CLOEXP      POOL6-1                 FALSE     seqCount           
#> 2 CLOEXP      POOL6-1                 FALSE     fragmentEstimate   
#> 3 PROJECT1100 ABX-LR-PL5-POOL14-1     FALSE     seqCount           
#> 4 PROJECT1100 ABX-LR-PL5-POOL14-1     FALSE     fragmentEstimate   
#> 5 PROJECT1100 ABX-LR-PL6-POOL15-1     FALSE     seqCount           
#> 6 PROJECT1100 ABX-LR-PL6-POOL15-1     FALSE     fragmentEstimate   
#> 7 PROJECT1101 ABY-LR-PL4-POOL54-2     FALSE     seqCount           
#> 8 PROJECT1101 ABY-LR-PL4-POOL54-2     FALSE     fragmentEstimate   
#>   Files_found                                                                   
#>   <fs::path>                                                                    
#> 1 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/VA/CLOEXP/quan…
#> 2 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/VA/CLOEXP/quan…
#> 3 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/PROJECT1100/qu…
#> 4 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/PROJECT1100/qu…
#> 5 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/PROJECT1100/qu…
#> 6 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/PROJECT1100/qu…
#> 7 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/PROJECT1101/qu…
#> 8 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpyB82OP/fs/PROJECT1101/qu…

cis <- CIS_grubbs(matrices)

options(op)