obtain a single integration matrix from individual quantification matrices.

Takes a list of integration matrices referring to different qunatification types and merges them in a single data frame that has multiple value columns, each renamed according to their quantification type of reference.

comparison_matrix(
  x,
  fragmentEstimate = "fragmentEstimate",
  seqCount = "seqCount",
  barcodeCount = "barcodeCount",
  cellCount = "cellCount",
  ShsCount = "ShsCount"
)

Arguments

x
A named list of integration matrices, ideally obtained via import_parallel_Vispa2Matrices_interactive or import_parallel_Vispa2Matrices_auto. Names must be quantification types.

fragmentEstimate
The name of the output column for fragment estimate values

seqCount
The name of the output column for sequence count values

barcodeCount
The name of the output column for barcode count values

cellCount
The name of the output column for cell count values

ShsCount
The name of the output column for Shs count values

Value

A tibble

See also

Examples

op <- options("ISAnalytics.widgets" = FALSE)
path <- system.file("extdata", "ex_association_file.tsv",
    package = "ISAnalytics"
)
root_pth <- system.file("extdata", "fs.zip", package = "ISAnalytics")
root <- unzip_file_system(root_pth, "fs")
matrices <- import_parallel_Vispa2Matrices_auto(
    association_file = path, root = root,
    quantification_type = c("fragmentEstimate", "seqCount"),
    matrix_type = "annotated", workers = 2, patterns = NULL,
    matching_opt = "ANY",
    dates_format = "dmy", multi_quant_matrix = FALSE
)

#> *** Association file import summary ***
#> ℹ For detailed report please set option 'ISAnalytics.widgets' to TRUE
#> * Parsing problems detected: TRUE
#> * Date parsing problems: TRUE
#> * Column problems detected: TRUE
#> * NAs found in important columns: TRUE
#> * File system alignment: no problems detected

#> *** Association file import summary ***
#> ℹ For detailed report please set option 'ISAnalytics.widgets' to TRUE
#> * Parsing problems detected: TRUE
#> * Date parsing problems: TRUE
#> * Column problems detected: TRUE
#> * NAs found in important columns: TRUE
#> * File system alignment: no problems detected

#> [1] "--- REPORT: FILES IMPORTED ---"
#> # A tibble: 8 x 5
#>   ProjectID   concatenatePoolIDSeqRun Quantification_type
#>   <chr>       <chr>                   <chr>              
#> 1 CLOEXP      POOL6-1                 fragmentEstimate   
#> 2 PROJECT1100 ABX-LR-PL5-POOL14-1     fragmentEstimate   
#> 3 PROJECT1100 ABX-LR-PL6-POOL15-1     fragmentEstimate   
#> 4 PROJECT1101 ABY-LR-PL4-POOL54-2     fragmentEstimate   
#> 5 CLOEXP      POOL6-1                 seqCount           
#> 6 PROJECT1100 ABX-LR-PL5-POOL14-1     seqCount           
#> 7 PROJECT1100 ABX-LR-PL6-POOL15-1     seqCount           
#> 8 PROJECT1101 ABY-LR-PL4-POOL54-2     seqCount           
#>   Files_chosen                                                                  
#>   <fs::path>                                                                    
#> 1 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpo5nkko/fs/VA/CLOEXP/quan…
#> 2 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpo5nkko/fs/PROJECT1100/qu…
#> 3 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpo5nkko/fs/PROJECT1100/qu…
#> 4 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpo5nkko/fs/PROJECT1101/qu…
#> 5 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpo5nkko/fs/VA/CLOEXP/quan…
#> 6 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpo5nkko/fs/PROJECT1100/qu…
#> 7 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpo5nkko/fs/PROJECT1100/qu…
#> 8 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpo5nkko/fs/PROJECT1101/qu…
#>   Imported
#>   <lgl>   
#> 1 TRUE    
#> 2 TRUE    
#> 3 TRUE    
#> 4 TRUE    
#> 5 TRUE    
#> 6 TRUE    
#> 7 TRUE    
#> 8 TRUE    
#> [1] "--- SUMMARY OF FILES CHOSEN FOR IMPORT ---"
#> # A tibble: 8 x 4
#>   ProjectID   concatenatePoolIDSeqRun Quantification_type
#>   <chr>       <chr>                   <chr>              
#> 1 CLOEXP      POOL6-1                 fragmentEstimate   
#> 2 CLOEXP      POOL6-1                 seqCount           
#> 3 PROJECT1100 ABX-LR-PL5-POOL14-1     fragmentEstimate   
#> 4 PROJECT1100 ABX-LR-PL5-POOL14-1     seqCount           
#> 5 PROJECT1100 ABX-LR-PL6-POOL15-1     fragmentEstimate   
#> 6 PROJECT1100 ABX-LR-PL6-POOL15-1     seqCount           
#> 7 PROJECT1101 ABY-LR-PL4-POOL54-2     fragmentEstimate   
#> 8 PROJECT1101 ABY-LR-PL4-POOL54-2     seqCount           
#>   Files_chosen                                                                  
#>   <fs::path>                                                                    
#> 1 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpo5nkko/fs/VA/CLOEXP/quan…
#> 2 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpo5nkko/fs/VA/CLOEXP/quan…
#> 3 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpo5nkko/fs/PROJECT1100/qu…
#> 4 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpo5nkko/fs/PROJECT1100/qu…
#> 5 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpo5nkko/fs/PROJECT1100/qu…
#> 6 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpo5nkko/fs/PROJECT1100/qu…
#> 7 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpo5nkko/fs/PROJECT1101/qu…
#> 8 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpo5nkko/fs/PROJECT1101/qu…
#> [1] "--- INTEGRATION MATRICES FOUND REPORT ---"
#> # A tibble: 8 x 5
#>   ProjectID   concatenatePoolIDSeqRun Anomalies Quantification_type
#>   <chr>       <chr>                   <lgl>     <chr>              
#> 1 CLOEXP      POOL6-1                 FALSE     fragmentEstimate   
#> 2 CLOEXP      POOL6-1                 FALSE     seqCount           
#> 3 PROJECT1100 ABX-LR-PL5-POOL14-1     FALSE     fragmentEstimate   
#> 4 PROJECT1100 ABX-LR-PL5-POOL14-1     FALSE     seqCount           
#> 5 PROJECT1100 ABX-LR-PL6-POOL15-1     FALSE     fragmentEstimate   
#> 6 PROJECT1100 ABX-LR-PL6-POOL15-1     FALSE     seqCount           
#> 7 PROJECT1101 ABY-LR-PL4-POOL54-2     FALSE     fragmentEstimate   
#> 8 PROJECT1101 ABY-LR-PL4-POOL54-2     FALSE     seqCount           
#>   Files_found                                                                   
#>   <fs::path>                                                                    
#> 1 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpo5nkko/fs/VA/CLOEXP/quan…
#> 2 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpo5nkko/fs/VA/CLOEXP/quan…
#> 3 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpo5nkko/fs/PROJECT1100/qu…
#> 4 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpo5nkko/fs/PROJECT1100/qu…
#> 5 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpo5nkko/fs/PROJECT1100/qu…
#> 6 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpo5nkko/fs/PROJECT1100/qu…
#> 7 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpo5nkko/fs/PROJECT1101/qu…
#> 8 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpo5nkko/fs/PROJECT1101/qu…

total_matrix <- comparison_matrix(matrices)
options(op)