Computes the abundance of every integration in the sample.

Abundance is obtained for every row by calculating the ratio between the single value and the total value for the sample.

compute_abundance(x, columns = "Value", percentage = TRUE)

Arguments

x
An integration matrix - aka a data frame that includes the mandatory_IS_vars() as columns

columns
A character vector of column names to process, must be numeric or integer columns

percentage
Add abundance as percentage?

Value

An integration matrix

Details

Abundance will be computed upon the user selected columns in the columns parameter. For each column a corresponding relative abundance column (and optionally a percentage abundance column) will be produced.

See also

Examples

path <- system.file("extdata", "ex_annotated_ISMatrix.tsv.xz",
    package = "ISAnalytics"
)
matrix <- import_single_Vispa2Matrix(path)
abundance <- compute_abundance(matrix)