Integrations cumulative count in time by sample

This function was deprecated in favour of a single function, please use cumulative_is instead.

Usage

cumulative_count_union(
  x,
  association_file = NULL,
  timepoint_column = "TimePoint",
  key = c("SubjectID", "CellMarker", "Tissue", "TimePoint"),
  include_tp_zero = FALSE,
  zero = "0000",
  aggregate = FALSE,
  ...
)

Arguments

x

A simple integration matrix or an aggregated matrix (see details)

association_file

NULL or the association file for x if aggregate is set to TRUE

timepoint_column

What is the name of the time point column?

key

The aggregation key - must always contain the timepoint_column

include_tp_zero

Include timepoint 0?

zero

How is 0 coded in the data frame?

aggregate

Should x be aggregated?

...

Additional parameters to pass to aggregate_values_by_key

Value

A data frame

Examples

data("integration_matrices", package = "ISAnalytics")
data("association_file", package = "ISAnalytics")
aggreg <- aggregate_values_by_key(
    x = integration_matrices,
    association_file = association_file,
    value_cols = c("seqCount", "fragmentEstimate")
)
cumulative_count <- cumulative_count_union(aggreg)
#> Warning: `cumulative_count_union()` was deprecated in ISAnalytics 1.5.4.
#> ℹ Please use `cumulative_is()` instead.
#> ℹ Use option `counts = TRUE`. Function will be likely dropped in the next
#>   release cycle
cumulative_count
#> $coordinates
#>       SubjectID CellMarker Tissue TimePoint chr integration_locus strand
#>    1:     PT001        MNC     BM        30   1           8464757      -
#>    2:     PT001        MNC     BM        30   1          16186297      -
#>    3:     PT001        MNC     BM        30   1          40689188      +
#>    4:     PT001        MNC     BM        30   1         157759338      -
#>    5:     PT001        MNC     BM        30   1         234596545      -
#>   ---                                                                   
#> 2371:     PT002        MNC     PB       360   5         176653985      +
#> 2372:     PT002        MNC     PB       360   6         147673451      -
#> 2373:     PT002        MNC     PB       360   7           5039809      -
#> 2374:     PT002        MNC     PB       360   7          17227687      +
#> 2375:     PT002        MNC     PB       360   9         127351995      -
#>       GeneName GeneStrand
#>    1:     RERE          -
#>    2:     SPEN          +
#>    3:      RLF          +
#>    4:    FCRL1          -
#>    5:   TARBP1          -
#>   ---                    
#> 2371:     NSD1          +
#> 2372:   STXBP5          +
#> 2373: RNF216P1          +
#> 2374:      AHR          +
#> 2375:    NR6A1          -
#> 
#> $counts
#>     SubjectID CellMarker Tissue TimePoint is_n_cumulative
#>  1:     PT001        MNC     BM        30              54
#>  2:     PT001        MNC     BM        60             147
#>  3:     PT001        MNC     BM        90             179
#>  4:     PT001        MNC     BM       180             240
#>  5:     PT001        MNC     BM       360             240
#>  6:     PT001        MNC     PB        30              28
#>  7:     PT001        MNC     PB        60              77
#>  8:     PT001        MNC     PB        90             104
#>  9:     PT001        MNC     PB       180             121
#> 10:     PT001        MNC     PB       360             121
#> 11:     PT002        MNC     BM        30              98
#> 12:     PT002        MNC     BM        60             126
#> 13:     PT002        MNC     BM        90             141
#> 14:     PT002        MNC     BM       180             184
#> 15:     PT002        MNC     BM       360             265
#> 16:     PT002        MNC     PB        30              15
#> 17:     PT002        MNC     PB        60              26
#> 18:     PT002        MNC     PB        90              38
#> 19:     PT002        MNC     PB       180              62
#> 20:     PT002        MNC     PB       360             109
#>