Expands integration matrix with the cumulative is union over time.

Given an input integration matrix that can be grouped over time, this function adds integrations in groups assuming that if an integration is observed at time point "t" then it is also observed in time point "t+1".

cumulative_is(
  x,
  key = c("SubjectID", "CellMarker", "Tissue", "TimePoint"),
  timepoint_col = "TimePoint",
  include_tp_zero = FALSE,
  keep_og_is = TRUE,
  expand = FALSE
)

Arguments

x
An integration matrix, ideally aggregated via aggregate_values_by_key()

key
The aggregation key used

timepoint_col
The name of the time point column

include_tp_zero
Should time point 0 be included?

keep_og_is
Keep original set of integrations as a separate column?

expand
If FALSE, for each group, the set of integration sites is returned in a separate column as a nested table, otherwise the resulting column is unnested.

Value

A data frame

See also

Examples

data("integration_matrices", package = "ISAnalytics")
data("association_file", package = "ISAnalytics")
aggreg <- aggregate_values_by_key(
    x = rlang::current_env()$integration_matrices,
    association_file = rlang::current_env()$association_file,
    value_cols = c("seqCount", "fragmentEstimate")
)

#> Error in aggregate_values_by_key(x = rlang::current_env()$integration_matrices,     association_file = rlang::current_env()$association_file,     value_cols = c("seqCount", "fragmentEstimate")): is.data.frame(x) || is.list(x) is not TRUE

cumulated_is <- cumulative_is(aggreg)

#> Error in is.data.frame(x): object 'aggreg' not found

cumulated_is

#> Error in eval(expr, envir, enclos): object 'cumulated_is' not found