Import Vispa2 stats given the aligned association file.

Imports all the Vispa2 stats files for each pool provided the association file has been aligned with the file system (see import_association_file).

Usage

import_Vispa2_stats(
  association_file,
  file_prefixes = default_iss_file_prefixes(),
  join_with_af = TRUE,
  pool_col = "concatenatePoolIDSeqRun",
  report_path = default_report_path()
)

Arguments

association_file

The file system aligned association file (contains columns with absolute paths to the 'iss' folder)

file_prefixes

A character vector with known file prefixes to match on file names. NOTE: the elements represent regular expressions. For defaults see default_iss_file_prefixes.

join_with_af

Logical, if TRUE the imported stats files will be merged with the association file, if false a single data frame holding only the stats will be returned.

pool_col

A single string. What is the name of the pool column used in the Vispa2 run? This will be used as a key to perform a join operation with the stats files POOL column.

report_path

The path where the report file should be saved. Can be a folder, a file or NULL if no report should be produced. Defaults to {user_home}/ISAnalytics_reports.

Value

A data frame

See also

Examples

fs_path <- system.file("extdata", "fs.zip", package = "ISAnalytics")
fs <- unzip_file_system(fs_path, "fs")
af_path <- system.file("extdata", "asso.file.tsv.gz",
    package = "ISAnalytics"
)
af <- import_association_file(af_path,
    root = fs,
    import_iss = FALSE,
    report_path = NULL
)
stats_files <- import_Vispa2_stats(af,
    join_with_af = FALSE,
    report_path = NULL
)
#>   |                                                                              |                                                                      |   0%  |                                                                              |==================                                                    |  25%  |                                                                              |===================================                                   |  50%  |                                                                              |====================================================                  |  75%  |                                                                              |======================================================================| 100%
#> 
head(stats_files)
#>         RUN_NAME     POOL       TAG PHIX_MAPPING PLASMID_MAPPED_BYPOOL
#> 1: PJ01|POOL01-1 POOL01-1 LTR75LC38     43586699               2256176
#> 2: PJ01|POOL01-1 POOL01-1 LTR53LC32     43586699               2256176
#> 3: PJ01|POOL01-1 POOL01-1 LTR83LC66     43586699               2256176
#> 4: PJ01|POOL01-1 POOL01-1 LTR27LC94     43586699               2256176
#> 5: PJ01|POOL01-1 POOL01-1 LTR69LC52     43586699               2256176
#> 6: PJ01|POOL01-1 POOL01-1  LTR37LC2     43586699               2256176
#>    BARCODE_MUX LTR_IDENTIFIED TRIMMING_FINAL_LTRLC LV_MAPPED BWA_MAPPED_OVERALL
#> 1:      645026         645026               630965    211757             402477
#> 2:      652208         652177               649044    303300             322086
#> 3:      451519         451512               449669    204810             227275
#> 4:      426500         426499               425666    185752             223915
#> 5:       18300          18300                18290      6962              10487
#> 6:      729327         729327               727219    318653             369117
#>    ISS_MAPPED_OVERALL RAW_READS QUALITY_PASSED ISS_MAPPED_PP
#> 1:             219452        NA             NA            NA
#> 2:             222646        NA             NA            NA
#> 3:             149385        NA             NA            NA
#> 4:             143283        NA             NA            NA
#> 5:               5907        NA             NA            NA
#> 6:             235640        NA             NA            NA