Import the association file from disk

Imports the association file and immediately performs a check on the file system starting from the root to assess the alignment between the two.

import_association_file(path, root)

Arguments

path
The path on disk to the association file.

root
The path on disk of the root folder of Vispa2 output. See details.

Value

A tibble with the contents of the association file plus a column containing the path in the file system for every project and pool if found.

Details

The import series of functions is designed to work in combination with the use of Vispa2 pipeline, please refer to this article for more details: VISPA2: A Scalable Pipeline for High-Throughput Identification and Annotation of Vector Integration Sites.
The pipeline automatically produces an hierarchical structure in the file system which follows this schema:

/root_folder

Optional intermediate folders

ProjectID
|_bam
|_bcmuxall
|_bed
|_iss
|_quality
|_report
|_quantification
*|___concatenatePoolIDSeqRun

For each ProjectID there may be several nested PoolIDs. The alignment function only looks for PoolIDs in the quantification folder, since it's the location of the matrices to import. For more details on how to properly use these functions, refer to the vignette - vignette("how_to_import_functions")

See also

Examples

path <- system.file("extdata", "ex_association_file.tsv",
    package = "ISAnalytics"
)
root_pth <- system.file("extdata", "fs.zip", package = "ISAnalytics")
root <- unzip_file_system(root_pth, "fs")
association_file <- import_association_file(path, root)