Import the association file from disk

Imports the association file and optionally performs a check on the file system starting from the root to assess the alignment between the two.

Usage

import_association_file(
  path,
  root = NULL,
  dates_format = "ymd",
  separator = "\t",
  filter_for = NULL,
  import_iss = FALSE,
  convert_tp = TRUE,
  report_path = default_report_path(),
  transformations = default_af_transform(convert_tp),
  tp_padding = lifecycle::deprecated(),
  ...
)

Arguments

path

The path on disk to the association file.

root

The path on disk of the root folder of VISPA2 output or NULL. See details.

dates_format

A single string indicating how dates should be parsed. Must be a value in: date_formats()

separator

The column separator used in the file

filter_for

A named list where names represent column names that must be filtered. For example: list(ProjectID = c("PROJECT1", "PROJECT2)) will filter the association file so that it contains only those rows for which the value of the column "ProjectID" is one of the specified values. If multiple columns are present in the list all filtering conditions are applied as a logical AND.

import_iss

Import VISPA2 pool stats and merge them with the association file? Logical value

convert_tp

Should be time points be converted into months and years? Logical value

report_path

The path where the report file should be saved. Can be a folder or NULL if no report should be produced. Defaults to {user_home}/ISAnalytics_reports.

transformations

Either NULL or a named list of purrr-style lambdas where names are column names the function should be applied to.

tp_padding

Deprecated. Use transformations instead.

...

Additional arguments to pass to import_Vispa2_stats

Value

The data frame containing metadata

Details

Transformations

Lambdas provided in input in the transformations argument, must be transformations, aka functions that take in input a vector and return a vector of the same length as the input.

If the transformation list contains column names that are not present in the data frame, they are simply ignored.

File system alignment

If the root argument is set to NULL no file system alignment is performed. This allows to import the basic file but it won't be possible to perform automated matrix and stats import. For more details see the "How to use import functions" vignette: vignette("workflow_start", package = "ISAnalytics")

Time point conversion

The time point conversion is based on the following logic, given TPD is the column containing the time point expressed in days and TPM and TPY are respectively the time points expressed as month and years

If TPD is NA --> NA (for both months and years)

TPM = 0, TPY = 0 if and only if TPD = 0 For conversion in months:

TPM = ceiling(TPD/30) if TPD < 30 otherwise TPM = round(TPD/30) For conversion in years:

TPY = ceiling(TPD/360)

Required tags

The function will explicitly check for the presence of these tags:

project_id

pool_id

tag_seq

subject

tissue

tp_days

cell_marker

pcr_replicate

vispa_concatenate

pcr_repl_id

proj_folder

The function will use all the available specifications contained in association_file_columns(TRUE) to read and parse the file. If the specifications contain columns with a type "date", the function will parse the generic date with the format in the dates_format argument.

See also

Examples

fs_path <- generate_default_folder_structure(type = "correct")
af <- import_association_file(fs_path$af,
    root = fs_path$root,
    report_path = NULL
)
head(af)
#>    ProjectID  FUSIONID PoolID TagSequence SubjectID VectorType VectorID
#> 1:      PJ01 ET#382.46 POOL01   LTR75LC38     PT001      lenti    GLOBE
#> 2:      PJ01 ET#381.40 POOL01   LTR53LC32     PT001      lenti    GLOBE
#> 3:      PJ01  ET#381.9 POOL01   LTR83LC66     PT001      lenti    GLOBE
#> 4:      PJ01 ET#381.71 POOL01   LTR27LC94     PT001      lenti    GLOBE
#> 5:      PJ01  ET#381.2 POOL01   LTR69LC52     PT001      lenti    GLOBE
#> 6:      PJ01 ET#382.28 POOL01    LTR37LC2     PT001      lenti    GLOBE
#>    ExperimentID Tissue TimePoint DNAFragmentation PCRMethod TagIDextended
#> 1:         <NA>     PB      0060            SONIC      SLiM     LTR75LC38
#> 2:         <NA>     BM      0180            SONIC      SLiM     LTR53LC32
#> 3:         <NA>     BM      0180            SONIC      SLiM     LTR83LC66
#> 4:         <NA>     BM      0180            SONIC      SLiM     LTR27LC94
#> 5:         <NA>     PB      0180            SONIC      SLiM     LTR69LC52
#> 6:         <NA>     BM      0060            SONIC      SLiM      LTR37LC2
#>    Keywords CellMarker      TagID NGSProvider NGSTechnology ConverrtedFilesDir
#> 1:     <NA>        MNC LTR75.LC38        <NA>         HiSeq               <NA>
#> 2:     <NA>        MNC LTR53.LC32        <NA>         HiSeq               <NA>
#> 3:     <NA>        MNC LTR83.LC66        <NA>         HiSeq               <NA>
#> 4:     <NA>        MNC LTR27.LC94        <NA>         HiSeq               <NA>
#> 5:     <NA>        MNC LTR69.LC52        <NA>         HiSeq               <NA>
#> 6:     <NA>        MNC  LTR37.LC2        <NA>         HiSeq               <NA>
#>    ConverrtedFilesName SourceFileFolder SourceFileNameR1 SourceFileNameR2
#> 1:                <NA>             <NA>             <NA>             <NA>
#> 2:                <NA>             <NA>             <NA>             <NA>
#> 3:                <NA>             <NA>             <NA>             <NA>
#> 4:                <NA>             <NA>             <NA>             <NA>
#> 5:                <NA>             <NA>             <NA>             <NA>
#> 6:                <NA>             <NA>             <NA>             <NA>
#>    DNAnumber ReplicateNumber DNAextractionDate DNAngUsed LinearPCRID
#> 1: PT001-103               3        2016-03-16    23.184        <NA>
#> 2:  PT001-81               2        2016-07-15   181.440        <NA>
#> 3:  PT001-81               1        2016-07-15   181.440        <NA>
#> 4:  PT001-81               3        2016-07-15   181.440        <NA>
#> 5:  PT001-74               1        2016-07-15    23.058        <NA>
#> 6: PT001-107               2        2016-03-16   171.360        <NA>
#>    LinearPCRDate SonicationDate LigationDate 1stExpoPCRID 1stExpoPCRDate
#> 1:          <NA>     2016-11-02   2016-11-02    ET#380.46     2016-11-02
#> 2:          <NA>     2016-11-02   2016-11-02    ET#379.40     2016-11-02
#> 3:          <NA>     2016-11-02   2016-11-02     ET#379.9     2016-11-02
#> 4:          <NA>     2016-11-02   2016-11-02    ET#379.71     2016-11-02
#> 5:          <NA>     2016-11-02   2016-11-02     ET#379.2     2016-11-02
#> 6:          <NA>     2016-11-02   2016-11-02    ET#380.28     2016-11-02
#>    2ndExpoID 2ndExpoDate FusionPrimerPCRID FusionPrimerPCRDate   PoolDate
#> 1:      <NA>        <NA>         ET#382.46          2016-11-03 2016-11-07
#> 2:      <NA>        <NA>         ET#381.40          2016-11-03 2016-11-07
#> 3:      <NA>        <NA>          ET#381.9          2016-11-03 2016-11-07
#> 4:      <NA>        <NA>         ET#381.71          2016-11-03 2016-11-07
#> 5:      <NA>        <NA>          ET#381.2          2016-11-03 2016-11-07
#> 6:      <NA>        <NA>         ET#382.28          2016-11-03 2016-11-07
#>    SequencingDate  VCN Genome SequencingRound Genotype TestGroup  MOI
#> 1:     2016-11-15 0.30   hg19               1     <NA>      <NA> <NA>
#> 2:     2016-11-15 0.27   hg19               1     <NA>      <NA> <NA>
#> 3:     2016-11-15 0.27   hg19               1     <NA>      <NA> <NA>
#> 4:     2016-11-15 0.27   hg19               1     <NA>      <NA> <NA>
#> 5:     2016-11-15 0.24   hg19               1     <NA>      <NA> <NA>
#> 6:     2016-11-15 0.42   hg19               1     <NA>      <NA> <NA>
#>    Engraftment Transduction Notes AddedField1 AddedField2 AddedField3
#> 1:          NA           NA  <NA>        <NA>        <NA>        <NA>
#> 2:          NA           NA  <NA>        <NA>        <NA>        <NA>
#> 3:          NA           NA  <NA>        <NA>        <NA>        <NA>
#> 4:          NA           NA  <NA>        <NA>        <NA>        <NA>
#> 5:          NA           NA  <NA>        <NA>        <NA>        <NA>
#> 6:          NA           NA  <NA>        <NA>        <NA>        <NA>
#>    AddedField4 concatenatePoolIDSeqRun AddedField6_RelativeBloodPercentage
#> 1:        <NA>                POOL01-1                                <NA>
#> 2:        <NA>                POOL01-1                                <NA>
#> 3:        <NA>                POOL01-1                                <NA>
#> 4:        <NA>                POOL01-1                                <NA>
#> 5:        <NA>                POOL01-1                                <NA>
#> 6:        <NA>                POOL01-1                                <NA>
#>    AddedField7_PurityTestFeasibility AddedField8_FacsSeparationPurity Kapa
#> 1:                              <NA>                             <NA>   NA
#> 2:                              <NA>                             <NA>   NA
#> 3:                              <NA>                             <NA>   NA
#> 4:                              <NA>                             <NA>   NA
#> 5:                              <NA>                             <NA>   NA
#> 6:                              <NA>                             <NA>   NA
#>    ulForPool
#> 1:        NA
#> 2:        NA
#> 3:        NA
#> 4:        NA
#> 5:        NA
#> 6:        NA
#>                                                 CompleteAmplificationID
#> 1: PJ01_POOL01_LTR75LC38_PT001_PT001-103_lenti_GLOBE_PB_1_SLiM_0060_MNC
#> 2:  PJ01_POOL01_LTR53LC32_PT001_PT001-81_lenti_GLOBE_BM_1_SLiM_0180_MNC
#> 3:  PJ01_POOL01_LTR83LC66_PT001_PT001-81_lenti_GLOBE_BM_1_SLiM_0180_MNC
#> 4:  PJ01_POOL01_LTR27LC94_PT001_PT001-81_lenti_GLOBE_BM_1_SLiM_0180_MNC
#> 5:  PJ01_POOL01_LTR69LC52_PT001_PT001-74_lenti_GLOBE_PB_1_SLiM_0180_MNC
#> 6:  PJ01_POOL01_LTR37LC2_PT001_PT001-107_lenti_GLOBE_BM_1_SLiM_0060_MNC
#>                  UniqueID StudyTestID StudyTestGroup MouseID Tigroup Tisource
#> 1: ID00000000000000007433        <NA>           <NA>    <NA>    <NA>     <NA>
#> 2: ID00000000000000007340        <NA>           <NA>    <NA>    <NA>     <NA>
#> 3: ID00000000000000007310        <NA>           <NA>    <NA>    <NA>     <NA>
#> 4: ID00000000000000007370        <NA>           <NA>    <NA>    <NA>     <NA>
#> 5: ID00000000000000007303        <NA>           <NA>    <NA>    <NA>     <NA>
#> 6: ID00000000000000007417        <NA>           <NA>    <NA>    <NA>     <NA>
#>    PathToFolderProjectID SamplesNameCheck TimepointDays TimepointMonths
#> 1:                 /PJ01             <NA>          0060              02
#> 2:                 /PJ01             <NA>          0180              06
#> 3:                 /PJ01             <NA>          0180              06
#> 4:                 /PJ01             <NA>          0180              06
#> 5:                 /PJ01             <NA>          0180              06
#> 6:                 /PJ01             <NA>          0060              02
#>    TimepointYears ng DNA corrected                    Path
#> 1:             01            23.18 /tmp/RtmpUrB7Q6/fs/PJ01
#> 2:             01           181.44 /tmp/RtmpUrB7Q6/fs/PJ01
#> 3:             01           181.44 /tmp/RtmpUrB7Q6/fs/PJ01
#> 4:             01           181.44 /tmp/RtmpUrB7Q6/fs/PJ01
#> 5:             01            23.06 /tmp/RtmpUrB7Q6/fs/PJ01
#> 6:             01           171.36 /tmp/RtmpUrB7Q6/fs/PJ01
#>                                         Path_quant
#> 1: /tmp/RtmpUrB7Q6/fs/PJ01/quantification/POOL01-1
#> 2: /tmp/RtmpUrB7Q6/fs/PJ01/quantification/POOL01-1
#> 3: /tmp/RtmpUrB7Q6/fs/PJ01/quantification/POOL01-1
#> 4: /tmp/RtmpUrB7Q6/fs/PJ01/quantification/POOL01-1
#> 5: /tmp/RtmpUrB7Q6/fs/PJ01/quantification/POOL01-1
#> 6: /tmp/RtmpUrB7Q6/fs/PJ01/quantification/POOL01-1
#>                                Path_iss
#> 1: /tmp/RtmpUrB7Q6/fs/PJ01/iss/POOL01-1
#> 2: /tmp/RtmpUrB7Q6/fs/PJ01/iss/POOL01-1
#> 3: /tmp/RtmpUrB7Q6/fs/PJ01/iss/POOL01-1
#> 4: /tmp/RtmpUrB7Q6/fs/PJ01/iss/POOL01-1
#> 5: /tmp/RtmpUrB7Q6/fs/PJ01/iss/POOL01-1
#> 6: /tmp/RtmpUrB7Q6/fs/PJ01/iss/POOL01-1