Import integration matrices from paths in the association file.

The function offers a convenient way of importing multiple integration matrices in an automated or semi-automated way. For more details see the "How to use import functions" vignette: vignette("workflow_start", package = "ISAnalytics")

Usage

import_parallel_Vispa2Matrices(
  association_file,
  quantification_type = c("seqCount", "fragmentEstimate"),
  matrix_type = c("annotated", "not_annotated"),
  workers = 2,
  multi_quant_matrix = TRUE,
  report_path = default_report_path(),
  patterns = NULL,
  matching_opt = matching_options(),
  mode = c("AUTO", "INTERACTIVE"),
  ...
)

Arguments

association_file

Data frame imported via import_association_file (with file system alignment) or a string containing the path to the association file on disk.

quantification_type

A vector of requested quantification_types. Possible choices are quantification_types

matrix_type

A single string representing the type of matrices to be imported. Can only be one in "annotated" or "not_annotated".

workers

A single integer representing the number of parallel workers to use for the import

multi_quant_matrix

If set to TRUE will produce a multi-quantification matrix through comparison_matrix instead of a list.

report_path

The path where the report file should be saved. Can be a folder or NULL if no report should be produced. Defaults to {user_home}/ISAnalytics_reports.

patterns

Relevant only if argument mode is set to AUTO. A character vector of additional patterns to match on file names. Please note that patterns must be regular expressions. Can be NULL if no patterns need to be matched.

matching_opt

Relevant only if argument mode is set to AUTO. A single value between matching_options

mode

A single value between AUTO and INTERACTIVE. If INTERACTIVE, the function will ask for input from the user on console, otherwise the process is fully automated (with limitations, see vignette).

...

<dynamic-dots> Additional named arguments to pass to ìmport_association_file, comparison_matrix and import_single_Vispa2_matrix

Value

Either a multi-quantification matrix or a list of integration matrices

Required tags

The function will explicitly check for the presence of these tags:

project_id

vispa_concatenate

See also

Examples

fs_path <- generate_default_folder_structure(type = "correct")
af <- import_association_file(fs_path$af,
    root = fs_path$root,
    report_path = NULL
)
matrices <- import_parallel_Vispa2Matrices(af,
    c("seqCount", "fragmentEstimate"),
    mode = "AUTO", report_path = NULL
)
head(matrices)
#>    chr integration_locus strand     GeneName GeneStrand
#> 1:  16          68164148      +       NFATC3          +
#> 2:   4         129390130      + LOC100507487          +
#> 3:   5          84009671      -        EDIL3          -
#> 4:  12          54635693      -         CBX5          -
#> 5:   2         181930711      +       UBE2E3          +
#> 6:  20          35920986      +       MANBAL          +
#>                                                 CompleteAmplificationID
#> 1: PJ01_POOL01_LTR75LC38_PT001_PT001-103_lenti_GLOBE_PB_1_SLiM_0060_MNC
#> 2: PJ01_POOL01_LTR75LC38_PT001_PT001-103_lenti_GLOBE_PB_1_SLiM_0060_MNC
#> 3: PJ01_POOL01_LTR75LC38_PT001_PT001-103_lenti_GLOBE_PB_1_SLiM_0060_MNC
#> 4: PJ01_POOL01_LTR75LC38_PT001_PT001-103_lenti_GLOBE_PB_1_SLiM_0060_MNC
#> 5: PJ01_POOL01_LTR75LC38_PT001_PT001-103_lenti_GLOBE_PB_1_SLiM_0060_MNC
#> 6: PJ01_POOL01_LTR75LC38_PT001_PT001-103_lenti_GLOBE_PB_1_SLiM_0060_MNC
#>    fragmentEstimate seqCount
#> 1:       102.945718      182
#> 2:         3.013590        4
#> 3:         5.033742        5
#> 4:         9.129338        9
#> 5:        50.502734       83
#> 6:        16.394422       39