Import a single integration matrix from file

This function allows to read and import an integration matrix produced as the output of Vispa2 pipeline and converts it to a tidy format.

Usage

import_single_Vispa2Matrix(
  path,
  to_exclude = NULL,
  keep_excluded = FALSE,
  separator = "\t"
)

Arguments

path

The path to the file on disk

to_exclude

Either NULL or a character vector of column names that should be ignored when importing

keep_excluded

Keep the columns in to_exclude as additional id columns?

separator

The column delimiter used, defaults to \t

Value

A data.table object in tidy format

Details

For more details see the "How to use import functions" vignette: vignette("import_functions_howto", package = "ISAnalytics")

See also

Examples

fs_path <- system.file("extdata", "fs.zip", package = "ISAnalytics")
fs <- unzip_file_system(fs_path, "fs")
matrix_path <- fs::path(
    fs,
    "PJ01",
    "quantification",
    "POOL01-1",
    "PJ01_POOL01-1_seqCount_matrix.no0.annotated.tsv.gz"
)
matrix <- import_single_Vispa2Matrix(matrix_path)
#> Reading file...
#> ℹ Mode: fread
#> *** File info *** 
#> * --- Matrix type: NEW
#> * --- Annotated: TRUE
#> * --- Dimensions: 261 x 29
#> * --- Read mode:  fread
#> Reshaping...
#>   |                                                                              |                                                                      |   0%  |                                                                              |===                                                                   |   4%  |                                                                              |======                                                                |   8%  |                                                                              |=========                                                             |  12%  |                                                                              |============                                                          |  17%  |                                                                              |===============                                                       |  21%  |                                                                              |==================                                                    |  25%  |                                                                              |====================                                                  |  29%  |                                                                              |=======================                                               |  33%  |                                                                              |==========================                                            |  38%  |                                                                              |=============================                                         |  42%  |                                                                              |================================                                      |  46%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================                                |  54%  |                                                                              |=========================================                             |  58%  |                                                                              |============================================                          |  62%  |                                                                              |===============================================                       |  67%  |                                                                              |==================================================                    |  71%  |                                                                              |====================================================                  |  75%  |                                                                              |=======================================================               |  79%  |                                                                              |==========================================================            |  83%  |                                                                              |=============================================================         |  88%  |                                                                              |================================================================      |  92%  |                                                                              |===================================================================   |  96%  |                                                                              |======================================================================| 100%
#> 
#> Done!
head(matrix)
#>    chr integration_locus strand     GeneName GeneStrand
#> 1:  16          68164148      +       NFATC3          +
#> 2:  16           1762026      -     MAPK8IP3          +
#> 3:  16          15966129      -        FOPNL          -
#> 4:  16           1762026      -     MAPK8IP3          +
#> 5:  16          29843197      -          MVP          +
#> 6:  16           1650839      + LOC105371046          +
#>                                                 CompleteAmplificationID Value
#> 1: PJ01_POOL01_LTR75LC38_PT001_PT001-103_lenti_GLOBE_PB_1_SLiM_0060_MNC   182
#> 2:  PJ01_POOL01_LTR53LC32_PT001_PT001-81_lenti_GLOBE_BM_1_SLiM_0180_MNC   727
#> 3:  PJ01_POOL01_LTR53LC32_PT001_PT001-81_lenti_GLOBE_BM_1_SLiM_0180_MNC   821
#> 4:  PJ01_POOL01_LTR83LC66_PT001_PT001-81_lenti_GLOBE_BM_1_SLiM_0180_MNC    37
#> 5:  PJ01_POOL01_LTR83LC66_PT001_PT001-81_lenti_GLOBE_BM_1_SLiM_0180_MNC   983
#> 6:  PJ01_POOL01_LTR83LC66_PT001_PT001-81_lenti_GLOBE_BM_1_SLiM_0180_MNC   389