Sharing of integration sites between given groups.

Computes the amount of integration sites shared between the groups identified in the input data.

Usage

is_sharing(
  ...,
  group_key = c("SubjectID", "CellMarker", "Tissue", "TimePoint"),
  group_keys = NULL,
  n_comp = 2,
  is_count = TRUE,
  relative_is_sharing = TRUE,
  minimal = TRUE,
  include_self_comp = FALSE,
  keep_genomic_coord = FALSE,
  table_for_venn = FALSE
)

Arguments

...

One or more integration matrices

group_key

Character vector of column names which identify a single group. An associated group id will be derived by concatenating the values of these fields, separated by "_"

group_keys

A list of keys for asymmetric grouping. If not NULL the argument group_key is ignored

n_comp

Number of comparisons to compute. This argument is relevant only if provided a single data frame and a single key.

is_count

Logical, if TRUE returns also the count of IS for each group and the count for the union set

relative_is_sharing

Logical, if TRUE also returns the relative sharing.

minimal

Compute only combinations instead of all possible permutations? If TRUE saves time and excludes redundant comparisons.

include_self_comp

Include comparisons with the same group?

keep_genomic_coord

If TRUE keeps the genomic coordinates of the shared integration sites in a dedicated column (as a nested table)

table_for_venn

Add column with truth tables for venn plots?

Value

A data frame

Details

An integration site is always identified by the triple (chr, integration_locus, strand), thus these columns must be present in the input(s).

The function accepts multiple inputs for different scenarios, please refer to the vignette vignette("sharing_analyses", package = "ISAnalytics") for a more in-depth explanation.

Output

The function outputs a single data frame containing all requested comparisons and optionally individual group counts, genomic coordinates of the shared integration sites and truth tables for plotting venn diagrams.

Plotting sharing

The sharing data obtained can be easily plotted in a heatmap via the function sharing_heatmap or via the function sharing_venn

See also

Examples

data("integration_matrices", package = "ISAnalytics")
data("association_file", package = "ISAnalytics")
aggreg <- aggregate_values_by_key(
    x = integration_matrices,
    association_file = association_file,
    value_cols = c("seqCount", "fragmentEstimate")
)
sharing <- is_sharing(aggreg)
#> Calculating combinations...
#> Done!
sharing
#>                     g1                g2 shared count_g1 count_g2 count_union
#>   1: PT001_MNC_BM_0030 PT001_MNC_BM_0060     21       54      114         147
#>   2: PT001_MNC_BM_0030 PT001_MNC_BM_0180     15       54       89         128
#>   3: PT001_MNC_BM_0060 PT001_MNC_BM_0180     13      114       89         190
#>   4: PT001_MNC_BM_0090 PT001_MNC_BM_0180     18       59       89         130
#>   5: PT001_MNC_BM_0030 PT001_MNC_BM_0360     21       54       78         111
#>  ---                                                                         
#> 186: PT001_MNC_BM_0090 PT002_MNC_PB_0090      0       59       24          83
#> 187: PT002_MNC_BM_0060 PT002_MNC_PB_0090      4       33       24          53
#> 188: PT002_MNC_BM_0090 PT002_MNC_PB_0090      2       40       24          62
#> 189: PT001_MNC_PB_0090 PT002_MNC_PB_0090      0       48       24          72
#> 190: PT001_MNC_PB_0360 PT002_MNC_PB_0090      0       43       24          67
#>         on_g1     on_g2  on_union
#>   1: 38.88889 18.421053 14.285714
#>   2: 27.77778 16.853933 11.718750
#>   3: 11.40351 14.606742  6.842105
#>   4: 30.50847 20.224719 13.846154
#>   5: 38.88889 26.923077 18.918919
#>  ---                             
#> 186:  0.00000  0.000000  0.000000
#> 187: 12.12121 16.666667  7.547170
#> 188:  5.00000  8.333333  3.225806
#> 189:  0.00000  0.000000  0.000000
#> 190:  0.00000  0.000000  0.000000