Sharing of integration sites between given groups.

Computes the amount of integrations shared between the groups identified by the fields in the group_key argument. An integration site is always identified by the triple (chr, integration_locus, strand), thus these columns must be present in the input data frame.

is_sharing(
  x,
  group_key = c("SubjectID", "CellMarker", "Tissue", "TimePoint"),
  is_count = TRUE,
  relative_is_sharing = TRUE
)

Arguments

x
An integration matrix, aka a data frame containing the columns mandatory_IS_vars(). See details.

group_key
Character vector of column names which identify a single group. An associated group id will be derived by concatenating the values of these fields, separated by "_"

is_count
Logical, if TRUE returns also the count of IS for each group

relative_is_sharing
Logical, if TRUE also returns the relative sharing. See details.

Value

A named list of data frames or a single data frame

Details

Input data frame

The data frame provided in input must be in a suitable format for the calculations to be accurate. Please note that this function does not perform any sort of aggregation, it only relies on counts of distinct integration sites.

Outputs

By default the function outputs a list of 2 data frames:

The classical sharing data frame with absolute values. If the argument relative_is_sharing is set to TRUE it also contains the relative sharing (see below).

The count of distinct IS for each group

The relative sharing is calculated, for each pair of groups (A, B) as 3 separate columns

Shared over A: (intersection(A,B) / |A|) * 100

Shared over B: (intersection(A,B) / |B|) * 100

Shared over union: (intersection(A,B) / |union(A,B)|) * 100

Plotting sharing

The sharing data obtained can be easily plotted in a heatmap via the function sharing_heatmap.

See also

Examples

data("integration_matrices", package = "ISAnalytics")
data("association_file", package = "ISAnalytics")
aggreg <- aggregate_values_by_key(
    x = integration_matrices,
    association_file = association_file,
    value_cols = c("seqCount", "fragmentEstimate")
)
sharing <- is_sharing(aggreg)
sharing

#> $is_count
#> # A tibble: 20 × 2
#>    group_id          num_IS
#>    <chr>              <int>
#>  1 PT001_MNC_BM_0030     54
#>  2 PT001_MNC_BM_0060    114
#>  3 PT001_MNC_BM_0180     89
#>  4 PT001_MNC_BM_0360     78
#>  5 PT002_MNC_BM_0360     91
#>  6 PT001_MNC_PB_0060     59
#>  7 PT001_MNC_PB_0180     29
#>  8 PT002_MNC_PB_0360     53
#>  9 PT002_MNC_PB_0180     32
#> 10 PT002_MNC_BM_0030     98
#> 11 PT002_MNC_PB_0060     18
#> 12 PT002_MNC_PB_0030     15
#> 13 PT001_MNC_PB_0030     28
#> 14 PT002_MNC_BM_0180     69
#> 15 PT001_MNC_BM_0090     59
#> 16 PT002_MNC_BM_0060     33
#> 17 PT002_MNC_BM_0090     40
#> 18 PT001_MNC_PB_0090     48
#> 19 PT001_MNC_PB_0360     43
#> 20 PT002_MNC_PB_0090     24
#> 
#> $sharing
#> # A tibble: 400 × 6
#>    group1            group2            shared  on_g1  on_g2 on_union
#>    <chr>             <chr>              <int>  <dbl>  <dbl>    <dbl>
#>  1 PT001_MNC_BM_0030 PT001_MNC_BM_0030     54 100    100     100    
#>  2 PT001_MNC_BM_0030 PT001_MNC_BM_0060     21  38.9   18.4    14.3  
#>  3 PT001_MNC_BM_0060 PT001_MNC_BM_0030     21  18.4   38.9    14.3  
#>  4 PT001_MNC_BM_0030 PT001_MNC_BM_0180     15  27.8   16.9    11.7  
#>  5 PT001_MNC_BM_0180 PT001_MNC_BM_0030     15  16.9   27.8    11.7  
#>  6 PT001_MNC_BM_0030 PT001_MNC_BM_0360     21  38.9   26.9    18.9  
#>  7 PT001_MNC_BM_0360 PT001_MNC_BM_0030     21  26.9   38.9    18.9  
#>  8 PT001_MNC_BM_0030 PT002_MNC_BM_0360      1   1.85   1.10    0.694
#>  9 PT002_MNC_BM_0360 PT001_MNC_BM_0030      1   1.10   1.85    0.694
#> 10 PT001_MNC_BM_0030 PT001_MNC_PB_0060      8  14.8   13.6     7.62 
#> # … with 390 more rows
#>