Identify and flag outliers based on pool fragments.

Identify and flag outliers based on expected number of raw reads per pool.

Usage

outliers_by_pool_fragments(
  metadata,
  key = "BARCODE_MUX",
  outlier_p_value_threshold = 0.01,
  normality_test = FALSE,
  normality_p_value_threshold = 0.05,
  transform_log2 = TRUE,
  per_pool_test = TRUE,
  pool_col = "PoolID",
  min_samples_per_pool = 5,
  flag_logic = "AND",
  keep_calc_cols = TRUE,
  report_path = default_report_path()
)

Arguments

metadata

The metadata data frame

key

A character vector of numeric column names

outlier_p_value_threshold

The p value threshold for a read to be considered an outlier

normality_test

Perform normality test? Normality is assessed for each column in the key using Shapiro-Wilk test and if the values do not follow a normal distribution, other calculations are skipped

normality_p_value_threshold

Normality threshold

transform_log2

Perform a log2 trasformation on values prior the actual calculations?

per_pool_test

Perform the test for each pool?

pool_col

A character vector of the names of the columns that uniquely identify a pool

min_samples_per_pool

The minimum number of samples that a pool needs to contain in order to be processed - relevant only if per_pool_test = TRUE

flag_logic

A character vector of logic operators to obtain a global flag formula - only relevant if the key is longer than one. All operators must be chosen between: AND, OR, XOR, NAND, NOR, XNOR

keep_calc_cols

Keep the calculation columns in the output data frame?

report_path

The path where the report file should be saved. Can be a folder, a file or NULL if no report should be produced. Defaults to {user_home}/ISAnalytics_reports.

Value

A data frame of metadata with the column to_remove

Details

Modular structure

The outlier filtering functions are structured in a modular fashion. There are 2 kind of functions:

Outlier tests - Functions that perform some kind of calculation based on inputs and flags metadata

Outlier filter - A function that takes one or more outlier tests, combines all the flags with a given logic and filters out rows that are flagged as outliers

This function is an outlier test, and calculates for each column in the key

The zscore of the values

The tstudent of the values

The the associated p-value (tdist)

Optionally the test can be performed for each pool and a normality test can be run prior the actual calculations. Samples are flagged if this condition is respected:

tdist < outlier_p_value_threshold & zscore < 0

If the key contains more than one column an additional flag logic can be specified for combining the results. Example: let's suppose the key contains the names of two columns, X and Y key = c("X", "Y") if we specify the the argument flag_logic = "AND" then the reads will be flagged based on this global condition: (tdist_X < outlier_p_value_threshold & zscore_X < 0) AND (tdist_Y < outlier_p_value_threshold & zscore_Y < 0)

The user can specify one or more logical operators that will be applied in sequence.

See also

Examples

data("association_file", package = "ISAnalytics")
flagged <- outliers_by_pool_fragments(association_file,
    report_path = NULL
)
#> Removing NAs from data...
#> Log2 transformation, removing values <= 0
head(flagged)
#>    ProjectID  FUSIONID PoolID TagSequence SubjectID VectorType VectorID
#> 1:      PJ01 ET#382.46 POOL01   LTR75LC38     PT001      lenti    GLOBE
#> 2:      PJ01 ET#381.40 POOL01   LTR53LC32     PT001      lenti    GLOBE
#> 3:      PJ01  ET#381.9 POOL01   LTR83LC66     PT001      lenti    GLOBE
#> 4:      PJ01 ET#381.71 POOL01   LTR27LC94     PT001      lenti    GLOBE
#> 5:      PJ01  ET#381.2 POOL01   LTR69LC52     PT001      lenti    GLOBE
#> 6:      PJ01 ET#382.28 POOL01    LTR37LC2     PT001      lenti    GLOBE
#>    ExperimentID Tissue TimePoint DNAFragmentation PCRMethod TagIDextended
#> 1:         <NA>     PB      0060            SONIC      SLiM     LTR75LC38
#> 2:         <NA>     BM      0180            SONIC      SLiM     LTR53LC32
#> 3:         <NA>     BM      0180            SONIC      SLiM     LTR83LC66
#> 4:         <NA>     BM      0180            SONIC      SLiM     LTR27LC94
#> 5:         <NA>     PB      0180            SONIC      SLiM     LTR69LC52
#> 6:         <NA>     BM      0060            SONIC      SLiM      LTR37LC2
#>    Keywords CellMarker      TagID NGSProvider NGSTechnology ConverrtedFilesDir
#> 1:     <NA>        MNC LTR75.LC38        <NA>         HiSeq               <NA>
#> 2:     <NA>        MNC LTR53.LC32        <NA>         HiSeq               <NA>
#> 3:     <NA>        MNC LTR83.LC66        <NA>         HiSeq               <NA>
#> 4:     <NA>        MNC LTR27.LC94        <NA>         HiSeq               <NA>
#> 5:     <NA>        MNC LTR69.LC52        <NA>         HiSeq               <NA>
#> 6:     <NA>        MNC  LTR37.LC2        <NA>         HiSeq               <NA>
#>    ConverrtedFilesName SourceFileFolder SourceFileNameR1 SourceFileNameR2
#> 1:                <NA>             <NA>             <NA>             <NA>
#> 2:                <NA>             <NA>             <NA>             <NA>
#> 3:                <NA>             <NA>             <NA>             <NA>
#> 4:                <NA>             <NA>             <NA>             <NA>
#> 5:                <NA>             <NA>             <NA>             <NA>
#> 6:                <NA>             <NA>             <NA>             <NA>
#>    DNAnumber ReplicateNumber DNAextractionDate DNAngUsed LinearPCRID
#> 1: PT001-103               3        2016-03-16    23.184        <NA>
#> 2:  PT001-81               2        2016-07-15   181.440        <NA>
#> 3:  PT001-81               1        2016-07-15   181.440        <NA>
#> 4:  PT001-81               3        2016-07-15   181.440        <NA>
#> 5:  PT001-74               1        2016-07-15    23.058        <NA>
#> 6: PT001-107               2        2016-03-16   171.360        <NA>
#>    LinearPCRDate SonicationDate LigationDate 1stExpoPCRID 1stExpoPCRDate
#> 1:          <NA>     2016-11-02   2016-11-02    ET#380.46     2016-11-02
#> 2:          <NA>     2016-11-02   2016-11-02    ET#379.40     2016-11-02
#> 3:          <NA>     2016-11-02   2016-11-02     ET#379.9     2016-11-02
#> 4:          <NA>     2016-11-02   2016-11-02    ET#379.71     2016-11-02
#> 5:          <NA>     2016-11-02   2016-11-02     ET#379.2     2016-11-02
#> 6:          <NA>     2016-11-02   2016-11-02    ET#380.28     2016-11-02
#>    2ndExpoID 2ndExpoDate FusionPrimerPCRID FusionPrimerPCRDate   PoolDate
#> 1:      <NA>        <NA>         ET#382.46          2016-11-03 2016-11-07
#> 2:      <NA>        <NA>         ET#381.40          2016-11-03 2016-11-07
#> 3:      <NA>        <NA>          ET#381.9          2016-11-03 2016-11-07
#> 4:      <NA>        <NA>         ET#381.71          2016-11-03 2016-11-07
#> 5:      <NA>        <NA>          ET#381.2          2016-11-03 2016-11-07
#> 6:      <NA>        <NA>         ET#382.28          2016-11-03 2016-11-07
#>    SequencingDate  VCN Genome SequencingRound Genotype TestGroup  MOI
#> 1:     2016-11-15 0.30   hg19               1     <NA>      <NA> <NA>
#> 2:     2016-11-15 0.27   hg19               1     <NA>      <NA> <NA>
#> 3:     2016-11-15 0.27   hg19               1     <NA>      <NA> <NA>
#> 4:     2016-11-15 0.27   hg19               1     <NA>      <NA> <NA>
#> 5:     2016-11-15 0.24   hg19               1     <NA>      <NA> <NA>
#> 6:     2016-11-15 0.42   hg19               1     <NA>      <NA> <NA>
#>    Engraftment Transduction Notes AddedField1 AddedField2 AddedField3
#> 1:          NA           NA  <NA>        <NA>        <NA>        <NA>
#> 2:          NA           NA  <NA>        <NA>        <NA>        <NA>
#> 3:          NA           NA  <NA>        <NA>        <NA>        <NA>
#> 4:          NA           NA  <NA>        <NA>        <NA>        <NA>
#> 5:          NA           NA  <NA>        <NA>        <NA>        <NA>
#> 6:          NA           NA  <NA>        <NA>        <NA>        <NA>
#>    AddedField4 concatenatePoolIDSeqRun AddedField6_RelativeBloodPercentage
#> 1:        <NA>                POOL01-1                                <NA>
#> 2:        <NA>                POOL01-1                                <NA>
#> 3:        <NA>                POOL01-1                                <NA>
#> 4:        <NA>                POOL01-1                                <NA>
#> 5:        <NA>                POOL01-1                                <NA>
#> 6:        <NA>                POOL01-1                                <NA>
#>    AddedField7_PurityTestFeasibility AddedField8_FacsSeparationPurity Kapa
#> 1:                                NA                               NA   NA
#> 2:                                NA                               NA   NA
#> 3:                                NA                               NA   NA
#> 4:                                NA                               NA   NA
#> 5:                                NA                               NA   NA
#> 6:                                NA                               NA   NA
#>    ulForPool
#> 1:        NA
#> 2:        NA
#> 3:        NA
#> 4:        NA
#> 5:        NA
#> 6:        NA
#>                                                 CompleteAmplificationID
#> 1: PJ01_POOL01_LTR75LC38_PT001_PT001-103_lenti_GLOBE_PB_1_SLiM_0060_MNC
#> 2:  PJ01_POOL01_LTR53LC32_PT001_PT001-81_lenti_GLOBE_BM_1_SLiM_0180_MNC
#> 3:  PJ01_POOL01_LTR83LC66_PT001_PT001-81_lenti_GLOBE_BM_1_SLiM_0180_MNC
#> 4:  PJ01_POOL01_LTR27LC94_PT001_PT001-81_lenti_GLOBE_BM_1_SLiM_0180_MNC
#> 5:  PJ01_POOL01_LTR69LC52_PT001_PT001-74_lenti_GLOBE_PB_1_SLiM_0180_MNC
#> 6:  PJ01_POOL01_LTR37LC2_PT001_PT001-107_lenti_GLOBE_BM_1_SLiM_0060_MNC
#>                  UniqueID StudyTestID StudyTestGroup MouseID Tigroup Tisource
#> 1: ID00000000000000007433        <NA>             NA      NA    <NA>     <NA>
#> 2: ID00000000000000007340        <NA>             NA      NA    <NA>     <NA>
#> 3: ID00000000000000007310        <NA>             NA      NA    <NA>     <NA>
#> 4: ID00000000000000007370        <NA>             NA      NA    <NA>     <NA>
#> 5: ID00000000000000007303        <NA>             NA      NA    <NA>     <NA>
#> 6: ID00000000000000007417        <NA>             NA      NA    <NA>     <NA>
#>    PathToFolderProjectID SamplesNameCheck TimepointDays TimepointMonths
#> 1:                 /PJ01             <NA>          0060              02
#> 2:                 /PJ01             <NA>          0180              06
#> 3:                 /PJ01             <NA>          0180              06
#> 4:                 /PJ01             <NA>          0180              06
#> 5:                 /PJ01             <NA>          0180              06
#> 6:                 /PJ01             <NA>          0060              02
#>    TimepointYears ng DNA corrected      RUN_NAME PHIX_MAPPING
#> 1:             01            23.18 PJ01|POOL01-1     43586699
#> 2:             01           181.44 PJ01|POOL01-1     43586699
#> 3:             01           181.44 PJ01|POOL01-1     43586699
#> 4:             01           181.44 PJ01|POOL01-1     43586699
#> 5:             01            23.06 PJ01|POOL01-1     43586699
#> 6:             01           171.36 PJ01|POOL01-1     43586699
#>    PLASMID_MAPPED_BYPOOL BARCODE_MUX LTR_IDENTIFIED TRIMMING_FINAL_LTRLC
#> 1:               2256176      645026         645026               630965
#> 2:               2256176      652208         652177               649044
#> 3:               2256176      451519         451512               449669
#> 4:               2256176      426500         426499               425666
#> 5:               2256176       18300          18300                18290
#> 6:               2256176      729327         729327               727219
#>    LV_MAPPED BWA_MAPPED_OVERALL ISS_MAPPED_OVERALL RAW_READS QUALITY_PASSED
#> 1:    211757             402477             219452        NA             NA
#> 2:    303300             322086             222646        NA             NA
#> 3:    204810             227275             149385        NA             NA
#> 4:    185752             223915             143283        NA             NA
#> 5:      6962              10487               5907        NA             NA
#> 6:    318653             369117             235640        NA             NA
#>    ISS_MAPPED_PP log2_BARCODE_MUX normality_BARCODE_MUX zscore_BARCODE_MUX
#> 1:            NA         19.29900                    NA          0.7338238
#> 2:            NA         19.31497                    NA          0.7411949
#> 3:            NA         18.78443                    NA          0.4963905
#> 4:            NA         18.70219                    NA          0.4584429
#> 5:            NA         14.15956                    NA         -1.6376186
#> 6:            NA         19.47621                    NA          0.8155914
#>    tstudent_BARCODE_MUX tdist_BARCODE_MUX processed flag_BARCODE_MUX to_remove
#> 1:            0.7422529        0.29678647      TRUE            FALSE     FALSE
#> 2:            0.7498983        0.29506532      TRUE            FALSE     FALSE
#> 3:            0.4987165        0.34660114      TRUE            FALSE     FALSE
#> 4:            0.4602087        0.35328569      TRUE            FALSE     FALSE
#> 5:            1.7457131        0.08872463      TRUE            FALSE     FALSE
#> 6:            0.8274007        0.27728811      TRUE            FALSE     FALSE