Re-aligns matrices of other quantification types based on the processed sequence count matrix.

This function should be used to keep data consistent among the same analysis: if for some reason you removed the collisions by passing only the sequence count matrix to the remove_collisions function, you should call this function afterwards, providing a list of other quantification matrices. NOTE: if you provided a list of several quantification types to remove_collisions before, there is no need to call this function.

realign_after_collisions(sc_matrix, other_matrices)

Arguments

sc_matrix
The sequence count matrix already processed for collisions via remove_collisions

other_matrices
A named list of matrices to re-align. Names in the list must be quantification types (quantification_types()) except "seqCount".

Value

A named list with re-aligned matrices

Details

For more details on how to use collision removal functionality: vignette("Collision removal functionality", package = "ISAnalytics")

See also

Examples

op <- options("ISAnalytics.widgets" = FALSE)
path <- system.file("extdata", "ex_association_file.tsv",
    package = "ISAnalytics"
)
root_pth <- system.file("extdata", "fs.zip", package = "ISAnalytics")
root <- unzip_file_system(root_pth, "fs")
association_file <- import_association_file(path, root,
    dates_format = "dmy"
)

#> *** Association file import summary ***
#> ℹ For detailed report please set option 'ISAnalytics.widgets' to TRUE
#> * Parsing problems detected: TRUE
#> * Date parsing problems: TRUE
#> * Column problems detected: TRUE
#> * NAs found in important columns: TRUE
#> * File system alignment: no problems detected

#> *** Association file import summary ***
#> ℹ For detailed report please set option 'ISAnalytics.widgets' to TRUE
#> * Parsing problems detected: TRUE
#> * Date parsing problems: TRUE
#> * Column problems detected: TRUE
#> * NAs found in important columns: TRUE
#> * File system alignment: no problems detected

matrices <- import_parallel_Vispa2Matrices_auto(
    association_file = association_file, root = NULL,
    quantification_type = c("fragmentEstimate", "seqCount"),
    matrix_type = "annotated", workers = 2,
    patterns = NULL, matching_opt = "ANY",
    multi_quant_matrix = FALSE
)

#> [1] "--- REPORT: FILES IMPORTED ---"
#> # A tibble: 8 x 5
#>   ProjectID   concatenatePoolIDSeqRun Quantification_type
#>   <chr>       <chr>                   <chr>              
#> 1 CLOEXP      POOL6-1                 fragmentEstimate   
#> 2 PROJECT1100 ABX-LR-PL5-POOL14-1     fragmentEstimate   
#> 3 PROJECT1100 ABX-LR-PL6-POOL15-1     fragmentEstimate   
#> 4 PROJECT1101 ABY-LR-PL4-POOL54-2     fragmentEstimate   
#> 5 CLOEXP      POOL6-1                 seqCount           
#> 6 PROJECT1100 ABX-LR-PL5-POOL14-1     seqCount           
#> 7 PROJECT1100 ABX-LR-PL6-POOL15-1     seqCount           
#> 8 PROJECT1101 ABY-LR-PL4-POOL54-2     seqCount           
#>   Files_chosen                                                                  
#>   <fs::path>                                                                    
#> 1 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpZ59MK9/fs/VA/CLOEXP/quan…
#> 2 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpZ59MK9/fs/PROJECT1100/qu…
#> 3 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpZ59MK9/fs/PROJECT1100/qu…
#> 4 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpZ59MK9/fs/PROJECT1101/qu…
#> 5 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpZ59MK9/fs/VA/CLOEXP/quan…
#> 6 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpZ59MK9/fs/PROJECT1100/qu…
#> 7 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpZ59MK9/fs/PROJECT1100/qu…
#> 8 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpZ59MK9/fs/PROJECT1101/qu…
#>   Imported
#>   <lgl>   
#> 1 TRUE    
#> 2 TRUE    
#> 3 TRUE    
#> 4 TRUE    
#> 5 TRUE    
#> 6 TRUE    
#> 7 TRUE    
#> 8 TRUE    
#> [1] "--- SUMMARY OF FILES CHOSEN FOR IMPORT ---"
#> # A tibble: 8 x 4
#>   ProjectID   concatenatePoolIDSeqRun Quantification_type
#>   <chr>       <chr>                   <chr>              
#> 1 CLOEXP      POOL6-1                 fragmentEstimate   
#> 2 CLOEXP      POOL6-1                 seqCount           
#> 3 PROJECT1100 ABX-LR-PL5-POOL14-1     fragmentEstimate   
#> 4 PROJECT1100 ABX-LR-PL5-POOL14-1     seqCount           
#> 5 PROJECT1100 ABX-LR-PL6-POOL15-1     fragmentEstimate   
#> 6 PROJECT1100 ABX-LR-PL6-POOL15-1     seqCount           
#> 7 PROJECT1101 ABY-LR-PL4-POOL54-2     fragmentEstimate   
#> 8 PROJECT1101 ABY-LR-PL4-POOL54-2     seqCount           
#>   Files_chosen                                                                  
#>   <fs::path>                                                                    
#> 1 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpZ59MK9/fs/VA/CLOEXP/quan…
#> 2 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpZ59MK9/fs/VA/CLOEXP/quan…
#> 3 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpZ59MK9/fs/PROJECT1100/qu…
#> 4 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpZ59MK9/fs/PROJECT1100/qu…
#> 5 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpZ59MK9/fs/PROJECT1100/qu…
#> 6 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpZ59MK9/fs/PROJECT1100/qu…
#> 7 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpZ59MK9/fs/PROJECT1101/qu…
#> 8 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpZ59MK9/fs/PROJECT1101/qu…
#> [1] "--- INTEGRATION MATRICES FOUND REPORT ---"
#> # A tibble: 8 x 5
#>   ProjectID   concatenatePoolIDSeqRun Anomalies Quantification_type
#>   <chr>       <chr>                   <lgl>     <chr>              
#> 1 CLOEXP      POOL6-1                 FALSE     fragmentEstimate   
#> 2 CLOEXP      POOL6-1                 FALSE     seqCount           
#> 3 PROJECT1100 ABX-LR-PL5-POOL14-1     FALSE     fragmentEstimate   
#> 4 PROJECT1100 ABX-LR-PL5-POOL14-1     FALSE     seqCount           
#> 5 PROJECT1100 ABX-LR-PL6-POOL15-1     FALSE     fragmentEstimate   
#> 6 PROJECT1100 ABX-LR-PL6-POOL15-1     FALSE     seqCount           
#> 7 PROJECT1101 ABY-LR-PL4-POOL54-2     FALSE     fragmentEstimate   
#> 8 PROJECT1101 ABY-LR-PL4-POOL54-2     FALSE     seqCount           
#>   Files_found                                                                   
#>   <fs::path>                                                                    
#> 1 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpZ59MK9/fs/VA/CLOEXP/quan…
#> 2 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpZ59MK9/fs/VA/CLOEXP/quan…
#> 3 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpZ59MK9/fs/PROJECT1100/qu…
#> 4 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpZ59MK9/fs/PROJECT1100/qu…
#> 5 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpZ59MK9/fs/PROJECT1100/qu…
#> 6 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpZ59MK9/fs/PROJECT1100/qu…
#> 7 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpZ59MK9/fs/PROJECT1101/qu…
#> 8 /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpZ59MK9/fs/PROJECT1101/qu…

sc_matrix <- remove_collisions(matrices$seqCount, association_file)

#> Found additional data relative to some projects that are not included in the imported matrices. Here is a summary

#> # A tibble: 1 x 3
#>   ProjectID   PoolID           
#>   <chr>       <chr>            
#> 1 PROJECT1101 ABY-LR-PL4-POOL54
#>   CompleteAmplificationID                                                       
#>   <chr>                                                                         
#> 1 PROJECT1101_ABY-LR-PL4-POOL54_LTR51LC51_VA2020-mix14_VA2020-mix14_lenti_737.p…

#> Identifying collisions...

#> Processing collisions...

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> 
#> [1] "--- SUMMARY ---"
#> # A tibble: 4 x 7
#>   SubjectID    count_integrations_before sumSeqReads_before
#>   <chr>                            <int>              <dbl>
#> 1 VA2020-mix10                      2092           10099664
#> 2 VA2020-mix11                      1332           15955128
#> 3 VA2020-mix12                       896            8172644
#> 4 VA2020-mix13                      1584           10035460
#>   count_integrations_after sumSeqReads_after delta_integrations delta_seqReads
#>                      <int>             <dbl>              <int>          <dbl>
#> 1                     1714           5052150                378        5047514
#> 2                      994           7980886                338        7974242
#> 3                      614           4088054                282        4084590
#> 4                     1258           5019402                326        5016058

#> Finished! You provided a single sequence count as matrix, to re-align other related matrices see ?realign_after_collisions

others <- matrices[!names(matrices) %in% "seqCount"]
aligned_matrices <- realign_after_collisions(sc_matrix, others)
options(op)