Re-aligns matrices of other quantification types based on the processed sequence count matrix.

This function should be used to keep data consistent among the same analysis: if for some reason you removed the collisions by passing only the sequence count matrix to remove_collisions(), you should call this function afterwards, providing a list of other quantification matrices. NOTE: if you provided a list of several quantification types to remove_collisions() before, there is no need to call this function.

Usage

realign_after_collisions(
  sc_matrix,
  other_matrices,
  sample_column = pcr_id_column()
)

Arguments

sc_matrix

The sequence count matrix already processed for collisions via remove_collisions()

other_matrices

A named list of matrices to re-align. Names in the list must be quantification types (quantification_types()) except "seqCount".

sample_column

The name of the column containing the sample identifier

Value

A named list with re-aligned matrices

Details

For more details on how to use collision removal functionality: vignette("workflow_start", package = "ISAnalytics")

See also

Examples

data("integration_matrices", package = "ISAnalytics")
data("association_file", package = "ISAnalytics")
separated <- separate_quant_matrices(
    integration_matrices
)
no_coll <- remove_collisions(
    x = separated$seqCount,
    association_file = association_file,
    quant_cols = c(seqCount = "Value"),
    report_path = NULL
)
#> Identifying collisions...
#> Processing collisions...
#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%
#> 
#> Finished!
realigned <- realign_after_collisions(
    sc_matrix = no_coll,
    other_matrices = list(fragmentEstimate = separated$fragmentEstimate)
)
realigned
#> $fragmentEstimate
#>       chr integration_locus strand     GeneName GeneStrand
#>    1:  16          68164148      +       NFATC3          +
#>    2:   4         129390130      + LOC100507487          +
#>    3:   5          84009671      -        EDIL3          -
#>    4:  12          54635693      -         CBX5          -
#>    5:   5          84009671      -        EDIL3          -
#>   ---                                                     
#> 1675:   6           3388625      -         HTR4          +
#> 1676:  16           3207754      +       UBE2D2          +
#> 1677:  19          13631664      -    LINC01133          +
#> 1678:  12          48119158      -        KMT2D          -
#> 1679:  14          30726412      -     PLEKHG4B          -
#>                                                    CompleteAmplificationID
#>    1: PJ01_POOL01_LTR75LC38_PT001_PT001-103_lenti_GLOBE_PB_1_SLiM_0060_MNC
#>    2:  PJ01_POOL01_LTR53LC32_PT001_PT001-81_lenti_GLOBE_BM_1_SLiM_0180_MNC
#>    3:  PJ01_POOL01_LTR53LC32_PT001_PT001-81_lenti_GLOBE_BM_1_SLiM_0180_MNC
#>    4:  PJ01_POOL01_LTR83LC66_PT001_PT001-81_lenti_GLOBE_BM_1_SLiM_0180_MNC
#>    5:  PJ01_POOL01_LTR83LC66_PT001_PT001-81_lenti_GLOBE_BM_1_SLiM_0180_MNC
#>   ---                                                                     
#> 1675: PJ01_POOL03_LTR93LC90_PT002_PT002-464_lenti_GLOBE_PB_1_SLiM_0360_MNC
#> 1676: PJ01_POOL03_LTR93LC90_PT002_PT002-464_lenti_GLOBE_PB_1_SLiM_0360_MNC
#> 1677: PJ01_POOL03_LTR93LC90_PT002_PT002-464_lenti_GLOBE_PB_1_SLiM_0360_MNC
#> 1678: PJ01_POOL03_LTR93LC90_PT002_PT002-464_lenti_GLOBE_PB_1_SLiM_0360_MNC
#> 1679: PJ01_POOL03_LTR93LC90_PT002_PT002-464_lenti_GLOBE_PB_1_SLiM_0360_MNC
#>            Value
#>    1: 102.945718
#>    2:  68.737467
#>    3:  67.123486
#>    4:  65.157600
#>    5:  61.469810
#>   ---           
#> 1675:  11.761600
#> 1676:  12.600475
#> 1677:   1.704548
#> 1678:  11.366729
#> 1679:   6.047534
#>