Identifies and removes collisions based on the sequence count matrix.

A collision is an integration (aka a unique combination of chr, integration_locus and strand) which is observed in more than one independent sample (a unique pair of ProjectID and SubjectID). The function tries to decide to which subject an integration should be assigned and if no decision can be taken, the integration is completely removed from the data frame.

remove_collisions(
  x,
  association_file,
  date_col = "SequencingDate",
  reads_ratio = 10,
  seq_count_col = "seqCount"
)

Arguments

x
A named list of matrices (names must be quantification types), a single integration matrix representing the sequence count matrix of interest or a multi-quantification matrix obtained via comparison_matrix

association_file
The association file imported via import_association_file

date_col
The date column that should be considered for the analysis. Must be one value in date_columns_coll()

reads_ratio
A single numeric value that represents the ratio that has to be considered when deciding between seqCount value.

seq_count_col
For support of multi-quantification matrix - the name of the sequence count values column

Value

A list of tibbles with removed collisions

Details

If you don't want the function to show details and messages do: options(ISAnalitics.verbose = FALSE). To restore to the original value: options(ISAnalitics.verbose = TRUE). For more details on how to use collision removal functionality: vignette("Collision removal functionality", package = "ISAnalytics")

See also

Examples

op <- options("ISAnalytics.widgets" = FALSE)
path <- system.file("extdata", "ex_association_file.tsv",
    package = "ISAnalytics"
)
root_pth <- system.file("extdata", "fs.zip", package = "ISAnalytics")
root <- unzip_file_system(root_pth, "fs")
association_file <- import_association_file(path, root)
matrices <- import_parallel_Vispa2Matrices_auto(
    association_file, NULL,
    c("fragmentEstimate", "seqCount"), "annotated", 2, NULL, "ANY"
)
matrices <- remove_collisions(matrices, association_file)

#> Found additional data relative to some projects that are not included in the imported matrices. Here is a summary

#> # A tibble: 1 x 3
#>   ProjectID   PoolID       CompleteAmplificationID                              
#>   <chr>       <chr>        <chr>                                                
#> 1 PROJECT1101 ABY-LR-PL4-… PROJECT1101_ABY-LR-PL4-POOL54_LTR51LC51_VA2020-mix14…

#> Identifying collisions...

#> Processing collisions...

#> Realigning matrices...

#> Finished!

options(op)