Identifies and removes collisions based on the sequence count matrix.

A collision is an integration (aka a unique combination of chr, integration_locus and strand) which is observed in more than one independent sample (a unique pair of ProjectID and SubjectID). The function tries to decide to which subject an integration should be assigned and if no decision can be taken, the integration is completely removed from the data frame.

remove_collisions(
  x,
  association_file,
  date_col = "SequencingDate",
  reads_ratio = 10,
  seq_count_col = "seqCount",
  max_rows_reports = 50
)

Arguments

x
A named list of matrices (names must be quantification types), a single integration matrix representing the sequence count matrix of interest or a multi-quantification matrix obtained via comparison_matrix

association_file
The association file imported via import_association_file

date_col
The date column that should be considered for the analysis. Must be one value in date_columns_coll()

reads_ratio
A single numeric value that represents the ratio that has to be considered when deciding between seqCount value.

seq_count_col
For support of multi-quantification matrix - the name of the sequence count values column

max_rows_reports
A numeric value, represents the maximum number of rows of the reports data frames that can be printed on console if the option ISAnalytics.verbose is active. If the data frames are too large they won't be printed on console - we recommend using widgets for detailed and more accessible info.

Value

A list of tibbles with removed collisions

Details

If you don't want the function to show details and messages do: options(ISAnalitics.verbose = FALSE). To restore to the original value: options(ISAnalitics.verbose = TRUE). For more details on how to use collision removal functionality: vignette("Collision removal functionality", package = "ISAnalytics")

See also

Examples

op <- options("ISAnalytics.widgets" = FALSE)
path <- system.file("extdata", "ex_association_file.tsv",
    package = "ISAnalytics"
)
root_pth <- system.file("extdata", "fs.zip", package = "ISAnalytics")
root <- unzip_file_system(root_pth, "fs")
association_file <- import_association_file(path, root,
    dates_format = "dmy")

#> # A tibble: 4 x 4
#>   ProjectID   concatenatePoolIDSeqRun Found
#>   <chr>       <chr>                   <lgl>
#> 1 CLOEXP      POOL6-1                 TRUE 
#> 2 PROJECT1100 ABX-LR-PL5-POOL14-1     TRUE 
#> 3 PROJECT1100 ABX-LR-PL6-POOL15-1     TRUE 
#> 4 PROJECT1101 ABY-LR-PL4-POOL54-2     TRUE 
#>   Path                                                             
#>   <chr>                                                            
#> 1 /tmp/RtmpaVHbny/fs/VA/CLOEXP/quantification/POOL6-1              
#> 2 /tmp/RtmpaVHbny/fs/PROJECT1100/quantification/ABX-LR-PL5-POOL14-1
#> 3 /tmp/RtmpaVHbny/fs/PROJECT1100/quantification/ABX-LR-PL6-POOL15-1
#> 4 /tmp/RtmpaVHbny/fs/PROJECT1101/quantification/ABY-LR-PL4-POOL54-2

matrices <- import_parallel_Vispa2Matrices_auto(
    association_file, NULL,
    c("fragmentEstimate", "seqCount"), "annotated", 2, NULL, "ANY"
)

#> [1] "--- REPORT: FILES IMPORTED ---"
#> # A tibble: 8 x 5
#>   ProjectID   concatenatePoolIDSeqRun Quantification_type
#>   <chr>       <chr>                   <chr>              
#> 1 CLOEXP      POOL6-1                 fragmentEstimate   
#> 2 PROJECT1100 ABX-LR-PL5-POOL14-1     fragmentEstimate   
#> 3 PROJECT1100 ABX-LR-PL6-POOL15-1     fragmentEstimate   
#> 4 PROJECT1101 ABY-LR-PL4-POOL54-2     fragmentEstimate   
#> 5 CLOEXP      POOL6-1                 seqCount           
#> 6 PROJECT1100 ABX-LR-PL5-POOL14-1     seqCount           
#> 7 PROJECT1100 ABX-LR-PL6-POOL15-1     seqCount           
#> 8 PROJECT1101 ABY-LR-PL4-POOL54-2     seqCount           
#>   Files_chosen                                                                  
#>   <fs::path>                                                                    
#> 1 /tmp/RtmpaVHbny/fs/VA/CLOEXP/quantification/POOL6-1/VA-CLOEXP-POOL6-1_fragmen…
#> 2 /tmp/RtmpaVHbny/fs/PROJECT1100/quantification/ABX-LR-PL5-POOL14-1/PROJECT1100…
#> 3 /tmp/RtmpaVHbny/fs/PROJECT1100/quantification/ABX-LR-PL6-POOL15-1/PROJECT1100…
#> 4 /tmp/RtmpaVHbny/fs/PROJECT1101/quantification/ABY-LR-PL4-POOL54-2/PROJECT1101…
#> 5 /tmp/RtmpaVHbny/fs/VA/CLOEXP/quantification/POOL6-1/VA-CLOEXP-POOL6-1_seqCoun…
#> 6 /tmp/RtmpaVHbny/fs/PROJECT1100/quantification/ABX-LR-PL5-POOL14-1/PROJECT1100…
#> 7 /tmp/RtmpaVHbny/fs/PROJECT1100/quantification/ABX-LR-PL6-POOL15-1/PROJECT1100…
#> 8 /tmp/RtmpaVHbny/fs/PROJECT1101/quantification/ABY-LR-PL4-POOL54-2/PROJECT1101…
#>   Imported
#>   <lgl>   
#> 1 TRUE    
#> 2 TRUE    
#> 3 TRUE    
#> 4 TRUE    
#> 5 TRUE    
#> 6 TRUE    
#> 7 TRUE    
#> 8 TRUE    
#> [1] "--- SUMMARY OF FILES CHOSEN FOR IMPORT ---"
#> # A tibble: 8 x 4
#>   ProjectID   concatenatePoolIDSeqRun Quantification_type
#>   <chr>       <chr>                   <chr>              
#> 1 CLOEXP      POOL6-1                 fragmentEstimate   
#> 2 CLOEXP      POOL6-1                 seqCount           
#> 3 PROJECT1100 ABX-LR-PL5-POOL14-1     fragmentEstimate   
#> 4 PROJECT1100 ABX-LR-PL5-POOL14-1     seqCount           
#> 5 PROJECT1100 ABX-LR-PL6-POOL15-1     fragmentEstimate   
#> 6 PROJECT1100 ABX-LR-PL6-POOL15-1     seqCount           
#> 7 PROJECT1101 ABY-LR-PL4-POOL54-2     fragmentEstimate   
#> 8 PROJECT1101 ABY-LR-PL4-POOL54-2     seqCount           
#>   Files_chosen                                                                  
#>   <fs::path>                                                                    
#> 1 /tmp/RtmpaVHbny/fs/VA/CLOEXP/quantification/POOL6-1/VA-CLOEXP-POOL6-1_fragmen…
#> 2 /tmp/RtmpaVHbny/fs/VA/CLOEXP/quantification/POOL6-1/VA-CLOEXP-POOL6-1_seqCoun…
#> 3 /tmp/RtmpaVHbny/fs/PROJECT1100/quantification/ABX-LR-PL5-POOL14-1/PROJECT1100…
#> 4 /tmp/RtmpaVHbny/fs/PROJECT1100/quantification/ABX-LR-PL5-POOL14-1/PROJECT1100…
#> 5 /tmp/RtmpaVHbny/fs/PROJECT1100/quantification/ABX-LR-PL6-POOL15-1/PROJECT1100…
#> 6 /tmp/RtmpaVHbny/fs/PROJECT1100/quantification/ABX-LR-PL6-POOL15-1/PROJECT1100…
#> 7 /tmp/RtmpaVHbny/fs/PROJECT1101/quantification/ABY-LR-PL4-POOL54-2/PROJECT1101…
#> 8 /tmp/RtmpaVHbny/fs/PROJECT1101/quantification/ABY-LR-PL4-POOL54-2/PROJECT1101…
#> [1] "--- INTEGRATION MATRICES FOUND REPORT ---"
#> # A tibble: 8 x 5
#>   ProjectID   concatenatePoolIDSeqRun Anomalies Quantification_type
#>   <chr>       <chr>                   <lgl>     <chr>              
#> 1 CLOEXP      POOL6-1                 FALSE     fragmentEstimate   
#> 2 CLOEXP      POOL6-1                 FALSE     seqCount           
#> 3 PROJECT1100 ABX-LR-PL5-POOL14-1     FALSE     fragmentEstimate   
#> 4 PROJECT1100 ABX-LR-PL5-POOL14-1     FALSE     seqCount           
#> 5 PROJECT1100 ABX-LR-PL6-POOL15-1     FALSE     fragmentEstimate   
#> 6 PROJECT1100 ABX-LR-PL6-POOL15-1     FALSE     seqCount           
#> 7 PROJECT1101 ABY-LR-PL4-POOL54-2     FALSE     fragmentEstimate   
#> 8 PROJECT1101 ABY-LR-PL4-POOL54-2     FALSE     seqCount           
#>   Files_found                                                                   
#>   <fs::path>                                                                    
#> 1 /tmp/RtmpaVHbny/fs/VA/CLOEXP/quantification/POOL6-1/VA-CLOEXP-POOL6-1_fragmen…
#> 2 /tmp/RtmpaVHbny/fs/VA/CLOEXP/quantification/POOL6-1/VA-CLOEXP-POOL6-1_seqCoun…
#> 3 /tmp/RtmpaVHbny/fs/PROJECT1100/quantification/ABX-LR-PL5-POOL14-1/PROJECT1100…
#> 4 /tmp/RtmpaVHbny/fs/PROJECT1100/quantification/ABX-LR-PL5-POOL14-1/PROJECT1100…
#> 5 /tmp/RtmpaVHbny/fs/PROJECT1100/quantification/ABX-LR-PL6-POOL15-1/PROJECT1100…
#> 6 /tmp/RtmpaVHbny/fs/PROJECT1100/quantification/ABX-LR-PL6-POOL15-1/PROJECT1100…
#> 7 /tmp/RtmpaVHbny/fs/PROJECT1101/quantification/ABY-LR-PL4-POOL54-2/PROJECT1101…
#> 8 /tmp/RtmpaVHbny/fs/PROJECT1101/quantification/ABY-LR-PL4-POOL54-2/PROJECT1101…

matrices <- remove_collisions(matrices, association_file)

#> Found additional data relative to some projects that are not included in the imported matrices. Here is a summary

#> # A tibble: 1 x 3
#>   ProjectID   PoolID           
#>   <chr>       <chr>            
#> 1 PROJECT1101 ABY-LR-PL4-POOL54
#>   CompleteAmplificationID                                                       
#>   <chr>                                                                         
#> 1 PROJECT1101_ABY-LR-PL4-POOL54_LTR51LC51_VA2020-mix14_VA2020-mix14_lenti_737.p…

#> Identifying collisions...

#> Processing collisions...

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> 
#> [1] "--- SUMMARY ---"
#> # A tibble: 4 x 7
#>   SubjectID    count_integrations_before sumSeqReads_before
#>   <chr>                            <int>              <int>
#> 1 VA2020-mix10                      2092           10099664
#> 2 VA2020-mix11                      1332           15955128
#> 3 VA2020-mix12                       896            8172644
#> 4 VA2020-mix13                      1584           10035460
#>   count_integrations_after sumSeqReads_after delta_integrations delta_seqReads
#>                      <int>             <int>              <int>          <int>
#> 1                     1714           5052150                378        5047514
#> 2                      994           7980886                338        7974242
#> 3                      614           4088054                282        4084590
#> 4                     1258           5019402                326        5016058

#> Realigning matrices...

#> Finished!

options(op)