Identifies and removes collisions.

A collision is an integration (aka a unique combination of the provided mandatory_IS_vars()) which is observed in more than one independent sample. The function tries to decide to which independent sample should an integration event be assigned to, and if no decision can be taken, the integration is completely removed from the data frame. For more details refer to the vignette "Collision removal functionality": vignette("workflow_start", package = "ISAnalytics")

Usage

remove_collisions(
  x,
  association_file,
  independent_sample_id = c("ProjectID", "SubjectID"),
  date_col = "SequencingDate",
  reads_ratio = 10,
  quant_cols = c(seqCount = "seqCount", fragmentEstimate = "fragmentEstimate"),
  report_path = default_report_path(),
  max_workers = NULL
)

Arguments

x

Either a multi-quantification matrix (recommended) or a named list of matrices (names must be quantification types)

association_file

The association file imported via import_association_file()

independent_sample_id

A character vector of column names that identify independent samples

date_col

The date column that should be considered.

reads_ratio

A single numeric value that represents the ratio that has to be considered when deciding between seqCount value.

quant_cols

A named character vector where names are quantification types and values are the names of the corresponding columns. The quantification seqCount MUST be included in the vector.

report_path

The path where the report file should be saved. Can be a folder or NULL if no report should be produced. Defaults to {user_home}/ISAnalytics_reports.

max_workers

Maximum number of parallel workers to distribute the workload. If NULL (default) produces the maximum amount of workers allowed, a numeric value is requested otherwise. WARNING: a higher number of workers speeds up computation at the cost of memory consumption! Tune this parameter accordingly.

Value

Either a multi-quantification matrix or a list of data frames

Required tags

The function will explicitly check for the presence of these tags:

project_id

pool_id

pcr_replicate

See also

Examples

data("integration_matrices", package = "ISAnalytics")
data("association_file", package = "ISAnalytics")
no_coll <- remove_collisions(
    x = integration_matrices,
    association_file = association_file,
    report_path = NULL
)
#> Identifying collisions...
#> Processing collisions...
#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%
#> 
#> Finished!
head(no_coll)
#>    chr integration_locus strand GeneName GeneStrand
#> 1:   1          27058748      -   ARID1A          +
#> 2:   1          27058748      -   ARID1A          +
#> 3:   1          27058748      -   ARID1A          +
#> 4:   1          27058748      -   ARID1A          +
#> 5:   1          45915674      -    TESK2          -
#> 6:   1          45915674      -    TESK2          -
#>                                                 CompleteAmplificationID
#> 1:  PJ01_POOL01_LTR37LC2_PT001_PT001-107_lenti_GLOBE_BM_1_SLiM_0060_MNC
#> 2: PJ01_POOL01_LTR83LC46_PT001_PT001-107_lenti_GLOBE_BM_1_SLiM_0060_MNC
#> 3: PJ01_POOL01_LTR85LC54_PT001_PT001-107_lenti_GLOBE_BM_1_SLiM_0060_MNC
#> 4: PJ01_POOL01_LTR77LC46_PT001_PT001-103_lenti_GLOBE_PB_1_SLiM_0060_MNC
#> 5:  PJ01_POOL01_LTR77LC56_PT001_PT001-93_lenti_GLOBE_PB_1_SLiM_0030_MNC
#> 6: PJ01_POOL01_LTR93LC56_PT001_PT001-112_lenti_GLOBE_PB_1_SLiM_0090_MNC
#>    seqCount fragmentEstimate
#> 1:     3583        13.197321
#> 2:     2978         8.058272
#> 3:     2006         7.046978
#> 4:        3         3.013293
#> 5:    15614         8.086461
#> 6:        4         1.000988