Identifies and removes collisions.

A collision is an integration (aka a unique combination of chr, integration_locus and strand) which is observed in more than one independent sample (a unique pair of ProjectID and SubjectID). The function tries to decide to which subject an integration should be assigned to and, if no decision can be taken, the integration is completely removed from the data frame. For more details refer to the vignette "Collision removal functionality": vignette("collision_removal", package = "ISAnalytics")

remove_collisions(
  x,
  association_file,
  date_col = "SequencingDate",
  reads_ratio = 10,
  quant_cols = c(seqCount = "seqCount", fragmentEstimate = "fragmentEstimate"),
  report_path = default_report_path()
)

Arguments

x
Either a multi-quantification matrix or a named list of matrices (names must be quantification types)

association_file
The association file imported via import_association_file()

date_col
The date column that should be considered. Must be one value in date_columns_coll()

reads_ratio
A single numeric value that represents the ratio that has to be considered when deciding between seqCount value.

quant_cols
A named character vector where names are quantification types and values are the names of the corresponding columns. The quantification seqCount MUST be included in the vector.

report_path
The path where the report file should be saved. Can be a folder, a file or NULL if no report should be produced. Defaults to {user_home}/ISAnalytics_reports.

Value

Either a multi-quantification matrix or a list of data frames

See also

Examples

data("integration_matrices", package = "ISAnalytics")
data("association_file", package = "ISAnalytics")
no_coll <- remove_collisions(
    x = integration_matrices,
    association_file = association_file,
    report_path = NULL
)

#> Identifying collisions...

#> Processing collisions...

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> 

#> Finished!

head(no_coll)

#> # A tibble: 6 × 8
#>   chr   integration_locus strand GeneName GeneStrand CompleteAmplifica… seqCount
#>   <chr>             <dbl> <chr>  <chr>    <chr>      <chr>                 <dbl>
#> 1 1              27058748 -      ARID1A   +          PJ01_POOL01_LTR37…     3583
#> 2 1              27058748 -      ARID1A   +          PJ01_POOL01_LTR83…     2978
#> 3 1              27058748 -      ARID1A   +          PJ01_POOL01_LTR85…     2006
#> 4 1              27058748 -      ARID1A   +          PJ01_POOL01_LTR77…        3
#> 5 1              45915674 -      TESK2    -          PJ01_POOL01_LTR77…    15614
#> 6 1              45915674 -      TESK2    -          PJ01_POOL01_LTR93…        4
#> # … with 1 more variable: fragmentEstimate <dbl>