Resets dynamic vars to the default values.

Reverts all changes to dynamic vars to the default values. For more details, refer to the dedicated vignette vignette("workflow_start", package="ISAnalytics").

reset_mandatory_IS_vars() re-sets the look-up table for mandatory IS vars.

reset_annotation_IS_vars() re-sets the look-up table for genomic annotation IS vars.

reset_af_columns_def() re-sets the look-up table for association file columns vars

reset_iss_stats_specs() re-sets the look-up table for VISPA2 pool statistics vars

reset_matrix_file_suffixes() re-sets the matrix file suffixes look-up table

reset_dyn_vars_config() re-sets all look-up tables

Usage

reset_mandatory_IS_vars()

reset_annotation_IS_vars()

reset_af_columns_def()

reset_iss_stats_specs()

reset_matrix_file_suffixes()

reset_dyn_vars_config()

Value

NULL

See also

Examples

reset_mandatory_IS_vars()
#> Mandatory IS vars reset to default

reset_annotation_IS_vars()
#> Annotation IS vars reset to default

reset_af_columns_def()
#> Association file columns specs reset to default

reset_iss_stats_specs()
#> ISS stats specs reset to default

reset_matrix_file_suffixes()
#> Matrix suffixes specs reset to default

reset_dyn_vars_config()
#> Mandatory IS vars reset to default
#> Annotation IS vars reset to default
#> Association file columns specs reset to default
#> ISS stats specs reset to default
#> Matrix suffixes specs reset to default