Separate a multiple-quantification matrix into single quantification matrices.

The function separates a single multi-quantification integration matrix, obtained via comparison_matrix, into single quantification matrices as a named list of tibbles.

Usage

separate_quant_matrices(
  x,
  fragmentEstimate = "fragmentEstimate",
  seqCount = "seqCount",
  barcodeCount = "barcodeCount",
  cellCount = "cellCount",
  ShsCount = "ShsCount",
  key = c(mandatory_IS_vars(), annotation_IS_vars(), "CompleteAmplificationID")
)

Arguments

x

Single integration matrix with multiple quantification value columns, likely obtained via comparison_matrix.

fragmentEstimate

Name of the fragment estimate values column in input

seqCount

Name of the sequence count values column in input

barcodeCount

Name of the barcode count values column in input

cellCount

Name of the cell count values column in input

ShsCount

Name of the shs count values column in input

key

Key columns to perform the joining operation

Value

A named list of tibbles, where names are quantification types

See also

Examples

data("integration_matrices", package = "ISAnalytics")
separated <- separate_quant_matrices(
    integration_matrices
)
separated
#> $fragmentEstimate
#>       chr integration_locus strand     GeneName GeneStrand
#>    1:  16          68164148      +       NFATC3          +
#>    2:   4         129390130      + LOC100507487          +
#>    3:   5          84009671      -        EDIL3          -
#>    4:  12          54635693      -         CBX5          -
#>    5:   5          84009671      -        EDIL3          -
#>   ---                                                     
#> 1685:   6           3388625      -         HTR4          +
#> 1686:  16           3207754      +       UBE2D2          +
#> 1687:  19          13631664      -    LINC01133          +
#> 1688:  12          48119158      -        KMT2D          -
#> 1689:  14          30726412      -     PLEKHG4B          -
#>                                                    CompleteAmplificationID
#>    1: PJ01_POOL01_LTR75LC38_PT001_PT001-103_lenti_GLOBE_PB_1_SLiM_0060_MNC
#>    2:  PJ01_POOL01_LTR53LC32_PT001_PT001-81_lenti_GLOBE_BM_1_SLiM_0180_MNC
#>    3:  PJ01_POOL01_LTR53LC32_PT001_PT001-81_lenti_GLOBE_BM_1_SLiM_0180_MNC
#>    4:  PJ01_POOL01_LTR83LC66_PT001_PT001-81_lenti_GLOBE_BM_1_SLiM_0180_MNC
#>    5:  PJ01_POOL01_LTR83LC66_PT001_PT001-81_lenti_GLOBE_BM_1_SLiM_0180_MNC
#>   ---                                                                     
#> 1685: PJ01_POOL03_LTR93LC90_PT002_PT002-464_lenti_GLOBE_PB_1_SLiM_0360_MNC
#> 1686: PJ01_POOL03_LTR93LC90_PT002_PT002-464_lenti_GLOBE_PB_1_SLiM_0360_MNC
#> 1687: PJ01_POOL03_LTR93LC90_PT002_PT002-464_lenti_GLOBE_PB_1_SLiM_0360_MNC
#> 1688: PJ01_POOL03_LTR93LC90_PT002_PT002-464_lenti_GLOBE_PB_1_SLiM_0360_MNC
#> 1689: PJ01_POOL03_LTR93LC90_PT002_PT002-464_lenti_GLOBE_PB_1_SLiM_0360_MNC
#>            Value
#>    1: 102.945718
#>    2:  68.737467
#>    3:  67.123486
#>    4:  65.157600
#>    5:  61.469810
#>   ---           
#> 1685:  11.761600
#> 1686:  12.600475
#> 1687:   1.704548
#> 1688:  11.366729
#> 1689:   6.047534
#> 
#> $seqCount
#>       chr integration_locus strand     GeneName GeneStrand
#>    1:  16          68164148      +       NFATC3          +
#>    2:   4         129390130      + LOC100507487          +
#>    3:   5          84009671      -        EDIL3          -
#>    4:  12          54635693      -         CBX5          -
#>    5:   5          84009671      -        EDIL3          -
#>   ---                                                     
#> 1685:   6           3388625      -         HTR4          +
#> 1686:  16           3207754      +       UBE2D2          +
#> 1687:  19          13631664      -    LINC01133          +
#> 1688:  12          48119158      -        KMT2D          -
#> 1689:  14          30726412      -     PLEKHG4B          -
#>                                                    CompleteAmplificationID
#>    1: PJ01_POOL01_LTR75LC38_PT001_PT001-103_lenti_GLOBE_PB_1_SLiM_0060_MNC
#>    2:  PJ01_POOL01_LTR53LC32_PT001_PT001-81_lenti_GLOBE_BM_1_SLiM_0180_MNC
#>    3:  PJ01_POOL01_LTR53LC32_PT001_PT001-81_lenti_GLOBE_BM_1_SLiM_0180_MNC
#>    4:  PJ01_POOL01_LTR83LC66_PT001_PT001-81_lenti_GLOBE_BM_1_SLiM_0180_MNC
#>    5:  PJ01_POOL01_LTR83LC66_PT001_PT001-81_lenti_GLOBE_BM_1_SLiM_0180_MNC
#>   ---                                                                     
#> 1685: PJ01_POOL03_LTR93LC90_PT002_PT002-464_lenti_GLOBE_PB_1_SLiM_0360_MNC
#> 1686: PJ01_POOL03_LTR93LC90_PT002_PT002-464_lenti_GLOBE_PB_1_SLiM_0360_MNC
#> 1687: PJ01_POOL03_LTR93LC90_PT002_PT002-464_lenti_GLOBE_PB_1_SLiM_0360_MNC
#> 1688: PJ01_POOL03_LTR93LC90_PT002_PT002-464_lenti_GLOBE_PB_1_SLiM_0360_MNC
#> 1689: PJ01_POOL03_LTR93LC90_PT002_PT002-464_lenti_GLOBE_PB_1_SLiM_0360_MNC
#>       Value
#>    1:   182
#>    2: 23219
#>    3: 20205
#>    4: 13269
#>    5: 14748
#>   ---      
#> 1685:   470
#> 1686:    50
#> 1687:   491
#> 1688:   340
#> 1689:    52
#>